@@ -3,12 +3,16 @@ module github.com/TileDB-Inc/TileDB-Go
 require (
 	github.com/mattn/go-pointer v0.0.1
 	github.com/stretchr/testify v1.6.1
+	go4.org/unsafe/assume-no-moving-gc v0.0.0-20230525183740-e7c30c78aeb2
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
-go 1.18
+// Local triggered panic when referencing enums
+retract v0.30.1
+
+go 1.20
