@@ -1,11 +1,7 @@
-//go:build !experimental
-// +build !experimental
-
 package main_test
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -112,7 +108,7 @@ func TestCantWrite(t *testing.T) {
 
 func read(t testing.TB, f string) string {
 	t.Helper()
-	in, err := ioutil.ReadFile(f)
+	in, err := os.ReadFile(f)
 	if err != nil {
 		t.Fatalf("error opening %v: %v", f, err)
 	}
