@@ -1,15 +1,6 @@
-//go:build experimental
-// +build experimental
-
-// This file declares Go bindings for experimental features in TileDB.
-// Experimental APIs to do not fall under the API compatibility guarantees and
-// might change between TileDB versions
-
 package tiledb
 
 /*
-#cgo LDFLAGS: -ltiledb
-#cgo linux LDFLAGS: -ldl
 #include <tiledb/tiledb_experimental.h>
 #include <stdlib.h>
 */
@@ -18,7 +9,6 @@ import "C"
 import (
 	"errors"
 	"fmt"
-	"runtime"
 	"unsafe"
 )
 
@@ -37,12 +27,7 @@ func NewArraySchemaEvolution(tdbCtx *Context) (*ArraySchemaEvolution, error) {
 		return nil, fmt.Errorf("error creating tiledb arraySchemaEvolution: %s",
 			arraySchemaEvolution.context.LastError())
 	}
-
-	// Set finalizer for free C pointer on gc
-	runtime.SetFinalizer(&arraySchemaEvolution,
-		func(arraySchemaEvolution *ArraySchemaEvolution) {
-			arraySchemaEvolution.Free()
-		})
+	freeOnGC(&arraySchemaEvolution)
 
 	return &arraySchemaEvolution, nil
 }
@@ -82,7 +67,7 @@ func (ase *ArraySchemaEvolution) AddAttribute(attribute *Attribute) error {
 	return nil
 }
 
-// DropAttribute drops an attribute to an array schema evolution
+// DropAttribute drops an attribute to an array schema evolution.
 func (ase *ArraySchemaEvolution) DropAttribute(name string) error {
 	cname := C.CString(name)
 	defer C.free(unsafe.Pointer(cname))
@@ -97,7 +82,7 @@ func (ase *ArraySchemaEvolution) DropAttribute(name string) error {
 	return nil
 }
 
-// Evolve evolves array schema of an array
+// Evolve evolves array schema of an array.
 func (ase *ArraySchemaEvolution) Evolve(uri string) error {
 	curi := C.CString(uri)
 	defer C.free(unsafe.Pointer(curi))
