@@ -1,19 +1,14 @@
 package tiledb
 
 /*
-#cgo LDFLAGS: -ltiledb
-#cgo linux LDFLAGS: -ldl
 #include <tiledb/tiledb.h>
 #include <stdlib.h>
 */
 import "C"
 
 import (
-	"errors"
 	"fmt"
 	"os"
-	"reflect"
-	"runtime"
 	"unsafe"
 )
 
@@ -22,16 +17,16 @@ Attribute describes an attribute of an Array cell.
 
 An attribute specifies a name and datatype for a particular value in each array cell. There are 3 supported attribute types:
 
-    Fundamental types, such as char, int, double, uint64, etc..
-    Fixed sized arrays: [N]T or make([]T, N), where T is a fundamental type
-    Variable length data: string, []T, where T is a fundamental type
+	Fundamental types, such as char, int, double, uint64, etc..
+	Fixed sized arrays: [N]T or make([]T, N), where T is a fundamental type
+	Variable length data: string, []T, where T is a fundamental type
 */
 type Attribute struct {
 	tiledbAttribute *C.tiledb_attribute_t
 	context         *Context
 }
 
-// NewAttribute alloc a new attribute
+// NewAttribute allocates a new attribute.
 func NewAttribute(context *Context, name string, datatype Datatype) (*Attribute, error) {
 	attribute := Attribute{context: context}
 	cname := C.CString(name)
@@ -41,11 +36,7 @@ func NewAttribute(context *Context, name string, datatype Datatype) (*Attribute,
 	if ret != C.TILEDB_OK {
 		return nil, fmt.Errorf("Error creating tiledb attribute: %s", context.LastError())
 	}
-
-	// Set finalizer for free C pointer on gc
-	runtime.SetFinalizer(&attribute, func(attribute *Attribute) {
-		attribute.Free()
-	})
+	freeOnGC(&attribute)
 
 	return &attribute, nil
 }
@@ -61,12 +52,12 @@ func (a *Attribute) Free() {
 	}
 }
 
-// Context exposes the internal TileDB context used to initialize the attribute
+// Context exposes the internal TileDB context used to initialize the attribute.
 func (a *Attribute) Context() *Context {
 	return a.context
 }
 
-// SetFilterList sets the attribute filterList
+// SetFilterList sets the attribute filterList.
 func (a *Attribute) SetFilterList(filterlist *FilterList) error {
 	ret := C.tiledb_attribute_set_filter_list(a.context.tiledbContext, a.tiledbAttribute, filterlist.tiledbFilterList)
 	if ret != C.TILEDB_OK {
@@ -75,23 +66,19 @@ func (a *Attribute) SetFilterList(filterlist *FilterList) error {
 	return nil
 }
 
-// FilterList returns a copy of the filter list for attribute
+// FilterList returns a copy of the filter list for attribute.
 func (a *Attribute) FilterList() (*FilterList, error) {
 	filterList := FilterList{context: a.context}
 	ret := C.tiledb_attribute_get_filter_list(a.context.tiledbContext, a.tiledbAttribute, &filterList.tiledbFilterList)
 	if ret != C.TILEDB_OK {
 		return nil, fmt.Errorf("Error getting tiledb attribute filter list: %s", a.context.LastError())
 	}
-
-	// Set finalizer for free C pointer on gc
-	runtime.SetFinalizer(&filterList, func(filterList *FilterList) {
-		filterList.Free()
-	})
+	freeOnGC(&filterList)
 
 	return &filterList, nil
 }
 
-// SetCellValNum Sets the number of attribute values per cell.
+// SetCellValNum sets the number of attribute values per cell.
 // This is inferred from the type parameter of the NewAttribute
 // function, but can also be set manually.
 func (a *Attribute) SetCellValNum(val uint32) error {
@@ -115,7 +102,7 @@ func (a *Attribute) CellValNum() (uint32, error) {
 	return uint32(cellValNum), nil
 }
 
-// CellSize gets attribute cell size
+// CellSize gets the attribute cell size.
 func (a *Attribute) CellSize() (uint64, error) {
 	var cellSize C.uint64_t
 	ret := C.tiledb_attribute_get_cell_size(a.context.tiledbContext, a.tiledbAttribute, &cellSize)
@@ -126,209 +113,172 @@ func (a *Attribute) CellSize() (uint64, error) {
 	return uint64(cellSize), nil
 }
 
-// SetFillValue Sets the default fill value for the input attribute. This value will
+// SetFillValue sets the default fill value for the input attribute. This value will
 // be used for the input attribute whenever querying (1) an empty cell in
 // a dense array, or (2) a non-empty cell (in either dense or sparse array)
 // when values on the input attribute are missing (e.g., if the user writes
 // a subset of the attributes in a write operation).
 // Applicable to var-sized attributes.
 // @note A call to `tiledb_attribute_cell_val_num` sets the fill value
-//      of the attribute to its default. Therefore, make sure you invoke
-//      `tiledb_attribute_set_fill_value` after deciding on the number
-//      of values this attribute will hold in each cell.
+//
+//	of the attribute to its default. Therefore, make sure you invoke
+//	`tiledb_attribute_set_fill_value` after deciding on the number
+//	of values this attribute will hold in each cell.
+//
 // @note For fixed-sized attributes, the input `size` should be equal
-//      to the cell size.
+//
+//	to the cell size.
 func (a *Attribute) SetFillValue(value interface{}) error {
-
-	if value == nil {
-		return errors.New("Unrecognized value type passed: Cannot be a nil")
+	switch value := value.(type) {
+	case int:
+		return attributeSetFillValue(a, value)
+	case int8:
+		return attributeSetFillValue(a, value)
+	case int16:
+		return attributeSetFillValue(a, value)
+	case int32:
+		return attributeSetFillValue(a, value)
+	case int64:
+		return attributeSetFillValue(a, value)
+	case uint:
+		return attributeSetFillValue(a, value)
+	case uint8:
+		return attributeSetFillValue(a, value)
+	case uint16:
+		return attributeSetFillValue(a, value)
+	case uint32:
+		return attributeSetFillValue(a, value)
+	case uint64:
+		return attributeSetFillValue(a, value)
+	case float32:
+		return attributeSetFillValue(a, value)
+	case float64:
+		return attributeSetFillValue(a, value)
+	case bool:
+		return attributeSetFillValue(a, value)
+	case string:
+		cValue := unsafe.Pointer(C.CString(value))
+		defer C.free(cValue)
+		return attributeSetFillValueInternal(a, cValue, uint64(len(value)))
 	}
+	return fmt.Errorf("unrecognized fill value type %T", value)
+}
 
-	if reflect.TypeOf(value).Kind() == reflect.Slice {
-		return errors.New("Unrecognized value type passed: Cannot be a slice")
-	}
-
-	valueType := reflect.TypeOf(value).Kind()
-
-	cellValNum, err := a.CellValNum()
+func attributeSetFillValue[T scalarType](a *Attribute, value T) error {
+	valNum, err := a.CellValNum()
 	if err != nil {
 		return err
 	}
-
-	attrDataType, err := a.Type()
+	dataType, err := a.Type()
 	if err != nil {
 		return err
 	}
-
-	var valueSize C.uint64_t
-	if cellValNum == TILEDB_VAR_NUM {
-		valueSize = C.uint64_t(reflect.TypeOf(value).Size())
-	} else {
-		valueSize = C.uint64_t(attrDataType.Size() * uint64(cellValNum))
+	valueSize := uint64(unsafe.Sizeof(value))
+	if valNum != TILEDB_VAR_NUM {
+		valueSize = dataType.Size() * uint64(valNum)
 	}
+	return attributeSetFillValueInternal(a, unsafe.Pointer(&value), valueSize)
+}
 
-	var ret C.int32_t
-	switch valueType {
-	case reflect.Int:
-		tmpValue := value.(int)
-		ret = C.tiledb_attribute_set_fill_value(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize)
-	case reflect.Int8:
-		tmpValue := value.(int8)
-		ret = C.tiledb_attribute_set_fill_value(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize)
-	case reflect.Int16:
-		tmpValue := value.(int16)
-		ret = C.tiledb_attribute_set_fill_value(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize)
-	case reflect.Int32:
-		tmpValue := value.(int32)
-		ret = C.tiledb_attribute_set_fill_value(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize)
-	case reflect.Int64:
-		tmpValue := value.(int64)
-		ret = C.tiledb_attribute_set_fill_value(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize)
-	case reflect.Uint:
-		tmpValue := value.(uint)
-		ret = C.tiledb_attribute_set_fill_value(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize)
-	case reflect.Uint8:
-		tmpValue := value.(uint8)
-		ret = C.tiledb_attribute_set_fill_value(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize)
-	case reflect.Uint16:
-		tmpValue := value.(uint16)
-		ret = C.tiledb_attribute_set_fill_value(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize)
-	case reflect.Uint32:
-		tmpValue := value.(uint32)
-		ret = C.tiledb_attribute_set_fill_value(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize)
-	case reflect.Uint64:
-		tmpValue := value.(uint64)
-		ret = C.tiledb_attribute_set_fill_value(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize)
-	case reflect.Float32:
-		tmpValue := value.(float32)
-		ret = C.tiledb_attribute_set_fill_value(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize)
-	case reflect.Float64:
-		tmpValue := value.(float64)
-		ret = C.tiledb_attribute_set_fill_value(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize)
-	case reflect.String:
-		stringValue := value.(string)
-		valueSize = C.uint64_t(len(stringValue))
-		cTmpValue := C.CString(stringValue)
-		defer C.free(unsafe.Pointer(cTmpValue))
-		if valueSize > 0 {
-			ret = C.tiledb_attribute_set_fill_value(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(cTmpValue), valueSize)
-		}
-	default:
-		valueInterfaceVal := reflect.ValueOf(value)
-		return fmt.Errorf("Unrecognized value type passed: %s", valueInterfaceVal.Kind().String())
-	}
-
+func attributeSetFillValueInternal(a *Attribute, value unsafe.Pointer, valueSize uint64) error {
+	ret := C.tiledb_attribute_set_fill_value(
+		a.context.tiledbContext,
+		a.tiledbAttribute,
+		value,
+		C.uint64_t(valueSize),
+	)
 	if ret != C.TILEDB_OK {
-		return fmt.Errorf("Error filling attribute value: %s", a.context.LastError())
+		return fmt.Errorf("could not set attribute fill value: %w", a.context.LastError())
 	}
-
 	return nil
 }
 
-// SetFillValueNullable Sets the default fill value for the input attribute. This value will
+// SetFillValueNullable sets the default fill value for the input attribute. This value will
 // be used for the input attribute whenever querying (1) an empty cell in
 // a dense array, or (2) a non-empty cell (in either dense or sparse array)
 // when values on the input attribute are missing (e.g., if the user writes
 // a subset of the attributes in a write operation).
 // Applicable to var-sized attributes.
 // @note A call to `tiledb_attribute_cell_val_num` sets the fill value
-//      of the attribute to its default. Therefore, make sure you invoke
-//      `tiledb_attribute_set_fill_value` after deciding on the number
-//      of values this attribute will hold in each cell.
+//
+//	of the attribute to its default. Therefore, make sure you invoke
+//	`tiledb_attribute_set_fill_value` after deciding on the number
+//	of values this attribute will hold in each cell.
+//
 // @note For fixed-sized attributes, the input `size` should be equal
-//      to the cell size.
+//
+//	to the cell size.
 func (a *Attribute) SetFillValueNullable(value interface{}, valid bool) error {
-
-	if value == nil {
-		return errors.New("Unrecognized value type passed: Cannot be a nil")
+	switch value := value.(type) {
+	case int:
+		return attributeSetFillValueNullable(a, value, valid)
+	case int8:
+		return attributeSetFillValueNullable(a, value, valid)
+	case int16:
+		return attributeSetFillValueNullable(a, value, valid)
+	case int32:
+		return attributeSetFillValueNullable(a, value, valid)
+	case int64:
+		return attributeSetFillValueNullable(a, value, valid)
+	case uint:
+		return attributeSetFillValueNullable(a, value, valid)
+	case uint8:
+		return attributeSetFillValueNullable(a, value, valid)
+	case uint16:
+		return attributeSetFillValueNullable(a, value, valid)
+	case uint32:
+		return attributeSetFillValueNullable(a, value, valid)
+	case uint64:
+		return attributeSetFillValueNullable(a, value, valid)
+	case float32:
+		return attributeSetFillValueNullable(a, value, valid)
+	case float64:
+		return attributeSetFillValueNullable(a, value, valid)
+	case bool:
+		return attributeSetFillValueNullable(a, value, valid)
+	case string:
+		cValue := unsafe.Pointer(C.CString(value))
+		defer C.free(cValue)
+		return attributeSetFillValueNullableInternal(a, cValue, uint64(len(value)), valid)
 	}
+	return fmt.Errorf("unrecognized fill value type %T", value)
+}
 
-	if reflect.TypeOf(value).Kind() == reflect.Slice {
-		return errors.New("Unrecognized value type passed: Cannot be a slice")
-	}
-
-	valueType := reflect.TypeOf(value).Kind()
-
-	cellValNum, err := a.CellValNum()
+func attributeSetFillValueNullable[T scalarType](a *Attribute, value T, valid bool) error {
+	valNum, err := a.CellValNum()
 	if err != nil {
 		return err
 	}
-
-	attrDataType, err := a.Type()
+	dataType, err := a.Type()
 	if err != nil {
 		return err
 	}
-
-	var valueSize C.uint64_t
-	if cellValNum == TILEDB_VAR_NUM {
-		valueSize = C.uint64_t(reflect.TypeOf(value).Size())
-	} else {
-		valueSize = C.uint64_t(attrDataType.Size() * uint64(cellValNum))
+	valueSize := uint64(unsafe.Sizeof(value))
+	if valNum != TILEDB_VAR_NUM {
+		valueSize = dataType.Size() * uint64(valNum)
 	}
+	return attributeSetFillValueNullableInternal(a, unsafe.Pointer(&value), valueSize, valid)
+}
 
-	var ret C.int32_t
-	var cvalid C.uint8_t
+func attributeSetFillValueNullableInternal(a *Attribute, value unsafe.Pointer, valueSize uint64, valid bool) error {
+	cValid := C.uint8_t(0)
 	if valid {
-		cvalid = 1
+		cValid = 1
 	}
-	switch valueType {
-	case reflect.Int:
-		tmpValue := value.(int)
-		ret = C.tiledb_attribute_set_fill_value_nullable(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize, cvalid)
-	case reflect.Int8:
-		tmpValue := value.(int8)
-		ret = C.tiledb_attribute_set_fill_value_nullable(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize, cvalid)
-	case reflect.Int16:
-		tmpValue := value.(int16)
-		ret = C.tiledb_attribute_set_fill_value_nullable(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize, cvalid)
-	case reflect.Int32:
-		tmpValue := value.(int32)
-		ret = C.tiledb_attribute_set_fill_value_nullable(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize, cvalid)
-	case reflect.Int64:
-		tmpValue := value.(int64)
-		ret = C.tiledb_attribute_set_fill_value_nullable(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize, cvalid)
-	case reflect.Uint:
-		tmpValue := value.(uint)
-		ret = C.tiledb_attribute_set_fill_value_nullable(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize, cvalid)
-	case reflect.Uint8:
-		tmpValue := value.(uint8)
-		ret = C.tiledb_attribute_set_fill_value_nullable(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize, cvalid)
-	case reflect.Uint16:
-		tmpValue := value.(uint16)
-		ret = C.tiledb_attribute_set_fill_value_nullable(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize, cvalid)
-	case reflect.Uint32:
-		tmpValue := value.(uint32)
-		ret = C.tiledb_attribute_set_fill_value_nullable(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize, cvalid)
-	case reflect.Uint64:
-		tmpValue := value.(uint64)
-		ret = C.tiledb_attribute_set_fill_value_nullable(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize, cvalid)
-	case reflect.Float32:
-		tmpValue := value.(float32)
-		ret = C.tiledb_attribute_set_fill_value_nullable(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize, cvalid)
-	case reflect.Float64:
-		tmpValue := value.(float64)
-		ret = C.tiledb_attribute_set_fill_value_nullable(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(&tmpValue), valueSize, cvalid)
-	case reflect.String:
-		stringValue := value.(string)
-		valueSize = C.uint64_t(len(stringValue))
-		cTmpValue := C.CString(stringValue)
-		defer C.free(unsafe.Pointer(cTmpValue))
-		if valueSize > 0 {
-			ret = C.tiledb_attribute_set_fill_value_nullable(a.context.tiledbContext, a.tiledbAttribute, unsafe.Pointer(cTmpValue), valueSize, cvalid)
-		}
-	default:
-		valueInterfaceVal := reflect.ValueOf(value)
-		return fmt.Errorf("Unrecognized value type passed: %s", valueInterfaceVal.Kind().String())
-	}
-
+	ret := C.tiledb_attribute_set_fill_value_nullable(
+		a.context.tiledbContext,
+		a.tiledbAttribute,
+		value,
+		C.uint64_t(valueSize),
+		cValid,
+	)
 	if ret != C.TILEDB_OK {
-		return fmt.Errorf("Error filling attribute value: %s", a.context.LastError())
+		return fmt.Errorf("could not set attribute fill value: %w", a.context.LastError())
 	}
-
 	return nil
 }
 
-// GetFillValue Gets the default fill value for the input attribute. This value will
+// GetFillValue gets the default fill value for the input attribute. This value will
 // be used for the input attribute whenever querying (1) an empty cell in
 // a dense array, or (2) a non-empty cell (in either dense or sparse array)
 // when values on the input attribute are missing (e.g., if the user writes
@@ -356,7 +306,7 @@ func (a *Attribute) GetFillValue() (interface{}, uint64, error) {
 	return value, uint64(fillValueSize), nil
 }
 
-// GetFillValueNullable Gets the default fill value for the input attribute. This value will
+// GetFillValueNullable gets the default fill value for the input attribute. This value will
 // be used for the input attribute whenever querying (1) an empty cell in
 // a dense array, or (2) a non-empty cell (in either dense or sparse array)
 // when values on the input attribute are missing (e.g., if the user writes
@@ -385,7 +335,7 @@ func (a *Attribute) GetFillValueNullable() (interface{}, uint64, bool, error) {
 	return value, uint64(fillValueSize), cvalid == 1, nil
 }
 
-// Name returns name of attribute
+// Name returns the name of the attribute.
 func (a *Attribute) Name() (string, error) {
 	var cName *C.char
 	ret := C.tiledb_attribute_get_name(a.context.tiledbContext, a.tiledbAttribute, &cName)
@@ -396,7 +346,7 @@ func (a *Attribute) Name() (string, error) {
 	return C.GoString(cName), nil
 }
 
-// Type returns the attribute datatype
+// Type returns the attribute datatype.
 func (a *Attribute) Type() (Datatype, error) {
 	var attrType C.tiledb_datatype_t
 	ret := C.tiledb_attribute_get_type(a.context.tiledbContext, a.tiledbAttribute, &attrType)
@@ -406,7 +356,7 @@ func (a *Attribute) Type() (Datatype, error) {
 	return Datatype(attrType), nil
 }
 
-// DumpSTDOUT Dumps the attribute in ASCII format to stdout
+// DumpSTDOUT dumps the attribute in ASCII format to stdout.
 func (a *Attribute) DumpSTDOUT() error {
 	ret := C.tiledb_attribute_dump(a.context.tiledbContext, a.tiledbAttribute, C.stdout)
 	if ret != C.TILEDB_OK {
@@ -415,7 +365,7 @@ func (a *Attribute) DumpSTDOUT() error {
 	return nil
 }
 
-// Dump Dumps the attribute in ASCII format in the selected output.
+// Dump dumps the attribute in ASCII format to the given path.
 func (a *Attribute) Dump(path string) error {
 
 	if _, err := os.Stat(path); err == nil {
@@ -442,7 +392,7 @@ func (a *Attribute) Dump(path string) error {
 	return nil
 }
 
-// SetNullable Sets if the attribute is nullable or not.
+// SetNullable sets if the attribute is nullable or not.
 func (a *Attribute) SetNullable(nullable bool) error {
 	var cNullable C.uint8_t
 	if nullable {
