@@ -1,15 +1,12 @@
 package tiledb
 
 /*
-#cgo LDFLAGS: -ltiledb
-#cgo linux LDFLAGS: -ldl
 #include <tiledb/tiledb.h>
 #include <stdlib.h>
 */
 import "C"
 import (
 	"fmt"
-	"runtime"
 )
 
 // BufferList A list of TileDB BufferList objects
@@ -26,11 +23,7 @@ func NewBufferList(context *Context) (*BufferList, error) {
 	if ret != C.TILEDB_OK {
 		return nil, fmt.Errorf("Error creating tiledb buffer list: %s", bufferList.context.LastError())
 	}
-
-	// Set finalizer for free C pointer on gc
-	runtime.SetFinalizer(&bufferList, func(bufferList *BufferList) {
-		bufferList.Free()
-	})
+	freeOnGC(&bufferList)
 
 	return &bufferList, nil
 }
@@ -46,12 +39,12 @@ func (b *BufferList) Free() {
 	}
 }
 
-// Context exposes the internal TileDB context used to initialize the buffer list
+// Context exposes the internal TileDB context used to initialize the buffer list.
 func (b *BufferList) Context() *Context {
 	return b.context
 }
 
-// NumBuffers returns number of buffers in the list
+// NumBuffers returns number of buffers in the list.
 func (b *BufferList) NumBuffers() (uint64, error) {
 	var numBuffers C.uint64_t
 	ret := C.tiledb_buffer_list_get_num_buffers(b.context.tiledbContext, b.tiledbBufferList, &numBuffers)
@@ -63,13 +56,10 @@ func (b *BufferList) NumBuffers() (uint64, error) {
 	return uint64(numBuffers), nil
 }
 
-// GetBuffer returns a Buffer at the given index in the list
+// GetBuffer returns a Buffer at the given index in the list.
 func (b *BufferList) GetBuffer(bufferIndex uint) (*Buffer, error) {
 	buffer := Buffer{context: b.context}
-	// Set finalizer for free C pointer on gc
-	runtime.SetFinalizer(&buffer, func(buffer *Buffer) {
-		buffer.Free()
-	})
+	freeOnGC(&buffer)
 
 	ret := C.tiledb_buffer_list_get_buffer(b.context.tiledbContext, b.tiledbBufferList, C.uint64_t(bufferIndex), &buffer.tiledbBuffer)
 	if ret != C.TILEDB_OK {
@@ -79,7 +69,7 @@ func (b *BufferList) GetBuffer(bufferIndex uint) (*Buffer, error) {
 	return &buffer, nil
 }
 
-// TotalSize returns total number of bytes in the buffers in the list
+// TotalSize returns the total number of bytes in the buffers in the list.
 func (b *BufferList) TotalSize() (uint64, error) {
 	var totalSize C.uint64_t
 	ret := C.tiledb_buffer_list_get_total_size(b.context.tiledbContext, b.tiledbBufferList, &totalSize)
@@ -91,13 +81,10 @@ func (b *BufferList) TotalSize() (uint64, error) {
 	return uint64(totalSize), nil
 }
 
-// Flatten copies and concatenates all buffers in the list into a new buffer
+// Flatten copies and concatenates all buffers in the list into a new buffer.
 func (b *BufferList) Flatten() (*Buffer, error) {
 	buffer := Buffer{context: b.context}
-	// Set finalizer for free C pointer on gc
-	runtime.SetFinalizer(&buffer, func(buffer *Buffer) {
-		buffer.Free()
-	})
+	freeOnGC(&buffer)
 
 	ret := C.tiledb_buffer_list_flatten(b.context.tiledbContext, b.tiledbBufferList, &buffer.tiledbBuffer)
 
