@@ -71,9 +71,9 @@ func writeVacuumSparseArray(dir string, buffD []int32, data []int32) {
 
 	err = query.SetLayout(tiledb.TILEDB_UNORDERED)
 	checkError(err)
-	_, err = query.SetBuffer("d", buffD)
+	_, err = query.SetDataBuffer("d", buffD)
 	checkError(err)
-	_, err = query.SetBuffer("a", data)
+	_, err = query.SetDataBuffer("a", data)
 	checkError(err)
 
 	// Perform the write
@@ -103,9 +103,16 @@ func readVacuumSparseArray(dir string) {
 	checkError(err)
 	defer query.Free()
 
+	// Prepare the subarray
+	subarray, err := array.NewSubarray()
+	checkError(err)
+	defer subarray.Free()
+
 	err = query.SetLayout(tiledb.TILEDB_UNORDERED)
 	checkError(err)
-	err = query.AddRange(0, int32(1), int32(3))
+	err = subarray.AddRange(0, tiledb.MakeRange[int32](1, 3))
+	checkError(err)
+	err = query.SetSubarray(subarray)
 	checkError(err)
 
 	size, err := query.EstResultSize("a")
@@ -118,9 +125,9 @@ func readVacuumSparseArray(dir string) {
 	fmt.Printf("Estimated query size in bytes for dimension 'd': %d\n", *size)
 	buffD := make([]int32, (*size)/bytesizes.Int32)
 
-	_, err = query.SetBuffer("d", buffD)
+	_, err = query.SetDataBuffer("d", buffD)
 	checkError(err)
-	_, err = query.SetBuffer("a", buffA)
+	_, err = query.SetDataBuffer("a", buffA)
 	checkError(err)
 
 	// Submit the query
@@ -172,6 +179,7 @@ func consolidateVacuum(dir string) {
 
 	ctx, err := tiledb.NewContext(nil)
 	checkError(err)
+	defer ctx.Free()
 
 	// Prepare the array for reading
 	array, err := tiledb.NewArray(ctx, dir)
@@ -187,8 +195,9 @@ func consolidateVacuum(dir string) {
 
 	config, err := tiledb.NewConfig()
 	checkError(err)
+	defer config.Free()
 
-	err = config.Set("sm.consolidation.buffer_size", "4")
+	err = config.Set("sm.consolidation.buffer_size", "8")
 	checkError(err)
 
 	err = array.Consolidate(config)
