@@ -63,7 +63,7 @@ func writeDensePaddingArray(dir string) {
 	checkError(err)
 	defer ctx.Free()
 
-	subarray := []int32{2, 3, 1, 2}
+	subArray := []int32{2, 3, 1, 2}
 
 	// Open the array for writing.
 	array, err := tiledb.NewArray(ctx, dir)
@@ -78,13 +78,19 @@ func writeDensePaddingArray(dir string) {
 	checkError(err)
 	defer query.Free()
 
+	subarray, err := array.NewSubarray()
+	checkError(err)
+	defer subarray.Free()
+
 	// First submission
 	data := []int32{1, 2, 3, 4}
 	err = query.SetLayout(tiledb.TILEDB_ROW_MAJOR)
 	checkError(err)
-	_, err = query.SetBuffer("a", data)
+	_, err = query.SetDataBuffer("a", data)
 	checkError(err)
-	err = query.SetSubArray(subarray)
+	err = subarray.SetSubArray(subArray)
+	checkError(err)
+	err = query.SetSubarray(subarray)
 	checkError(err)
 
 	// Perform the write and close the array.
@@ -120,11 +126,18 @@ func readDensePaddingArray(dir string) {
 	checkError(err)
 	defer query.Free()
 
-	err = query.SetSubArray(subArray)
+	// Prepare the subarray
+	subarray, err := array.NewSubarray()
+	checkError(err)
+	defer subarray.Free()
+
+	err = subarray.SetSubArray(subArray)
+	checkError(err)
+	err = query.SetSubarray(subarray)
 	checkError(err)
 	err = query.SetLayout(tiledb.TILEDB_ROW_MAJOR)
 	checkError(err)
-	_, err = query.SetBuffer("a", data)
+	_, err = query.SetDataBuffer("a", data)
 	checkError(err)
 
 	// Submit the query and close the array.
