@@ -47,11 +47,15 @@ func testQueryConditionInt32(t *testing.T, array *Array) {
 		name           string
 		opValue        int32
 		op             QueryConditionOp
+		negate         bool
 		expectedValues []int32
 	}{
-		{"GreaterThan1", 1, TILEDB_QUERY_CONDITION_GT, []int32{2, 3}},
-		{"LessThan3", 3, TILEDB_QUERY_CONDITION_LT, []int32{1, 2}},
-		{"EqualTo2", 2, TILEDB_QUERY_CONDITION_EQ, []int32{2}},
+		{"GreaterThan1", 1, TILEDB_QUERY_CONDITION_GT, false, []int32{2, 3}},
+		{"NotGreaterThan1", 1, TILEDB_QUERY_CONDITION_GT, true, []int32{1}},
+		{"LessThan3", 3, TILEDB_QUERY_CONDITION_LT, false, []int32{1, 2}},
+		{"NotLessThan3", 3, TILEDB_QUERY_CONDITION_LT, true, []int32{3}},
+		{"EqualTo2", 2, TILEDB_QUERY_CONDITION_EQ, false, []int32{2}},
+		{"NotEqualTo2", 2, TILEDB_QUERY_CONDITION_EQ, true, []int32{1, 3}},
 	}
 	for _, c := range a1Cases {
 		c := c
@@ -62,12 +66,17 @@ func testQueryConditionInt32(t *testing.T, array *Array) {
 			require.NoError(t, err)
 			assert.NotNil(t, query)
 
-			_, err = query.SetBuffer("a1", a1DataRead)
+			_, err = query.SetDataBuffer("a1", a1DataRead)
 			require.NoError(t, err)
 			assert.NotNil(t, query)
 
 			qc, err := NewQueryCondition(array.context, "a1", c.op, c.opValue)
 			require.NoError(t, err)
+
+			if c.negate {
+				qc, err = NewQueryConditionNegated(array.context, qc)
+				require.NoError(t, err)
+			}
 
 			err = query.SetQueryCondition(qc)
 			require.NoError(t, err)
@@ -96,6 +105,7 @@ func testQueryConditionInt32(t *testing.T, array *Array) {
 		op2                  QueryConditionOp
 		op2Value             int32
 		combinationCondition QueryConditionCombinationOp
+		negate               bool
 		expectedValues       []int32
 	}{
 		{
@@ -105,7 +115,18 @@ func testQueryConditionInt32(t *testing.T, array *Array) {
 			op2:                  TILEDB_QUERY_CONDITION_LT,
 			op2Value:             3,
 			combinationCondition: TILEDB_QUERY_CONDITION_AND,
+			negate:               false,
 			expectedValues:       []int32{2},
+		},
+		{
+			name:                 "NotGreaterThan1AndLessThan3",
+			op1:                  TILEDB_QUERY_CONDITION_GT,
+			op1Value:             1,
+			op2:                  TILEDB_QUERY_CONDITION_LT,
+			op2Value:             3,
+			combinationCondition: TILEDB_QUERY_CONDITION_AND,
+			negate:               true,
+			expectedValues:       []int32{1, 3},
 		},
 	}
 	for _, c := range a1CombinationCases {
@@ -117,7 +138,7 @@ func testQueryConditionInt32(t *testing.T, array *Array) {
 			require.NoError(t, err)
 			assert.NotNil(t, query)
 
-			_, err = query.SetBuffer("a1", a1DataRead)
+			_, err = query.SetDataBuffer("a1", a1DataRead)
 			require.NoError(t, err)
 			assert.NotNil(t, query)
 
@@ -129,6 +150,11 @@ func testQueryConditionInt32(t *testing.T, array *Array) {
 
 			qc, err := NewQueryConditionCombination(array.context, qc1, c.combinationCondition, qc2)
 			require.NoError(t, err)
+
+			if c.negate {
+				qc, err = NewQueryConditionNegated(array.context, qc)
+				require.NoError(t, err)
+			}
 
 			err = query.SetQueryCondition(qc)
 			require.NoError(t, err)
@@ -170,7 +196,7 @@ func testQueryConditionTime(t *testing.T, array *Array) {
 			require.NoError(t, err)
 			assert.NotNil(t, query)
 
-			_, err = query.SetBuffer("a3", a3DataRead)
+			_, err = query.SetDataBuffer("a3", a3DataRead)
 			require.NoError(t, err)
 			assert.NotNil(t, query)
 			qc, err := NewQueryCondition(array.context, "a3", c.op, c.opValue)
@@ -218,7 +244,9 @@ func testQueryConditionBytes(t *testing.T, array *Array) {
 			require.NoError(t, err)
 			assert.NotNil(t, query)
 
-			_, _, err = query.SetBufferVar("a2", a2OffsetRead, a2DataRead)
+			_, err = query.SetDataBuffer("a2", a2DataRead)
+			require.NoError(t, err)
+			_, err = query.SetOffsetsBuffer("a2", a2OffsetRead)
 			require.NoError(t, err)
 			assert.NotNil(t, query)
 			qc, err := NewQueryCondition(array.context, "a2", c.op, c.opValue)
@@ -299,19 +327,22 @@ func createBasicTestArray(t testing.TB, identifier string) (*Array, error) {
 	if err := query.SetLayout(TILEDB_UNORDERED); err != nil {
 		return nil, err
 	}
-	if _, err = query.SetBuffer("a1", testAttributeValues.Attribute1); err != nil {
+	if _, err = query.SetDataBuffer("a1", testAttributeValues.Attribute1); err != nil {
 		return nil, err
 	}
-	if _, _, err = query.SetBufferVar("a2", testAttributeValues.Attribute2Offset, testAttributeValues.Attribute2); err != nil {
+	if _, err = query.SetDataBuffer("a2", testAttributeValues.Attribute2); err != nil {
 		return nil, err
 	}
-	if _, err = query.SetBuffer("a3", testAttributeValues.Attribute3); err != nil {
+	if _, err = query.SetOffsetsBuffer("a2", testAttributeValues.Attribute2Offset); err != nil {
 		return nil, err
 	}
-	if _, err := query.SetBuffer("rows", buffD1); err != nil {
+	if _, err = query.SetDataBuffer("a3", testAttributeValues.Attribute3); err != nil {
 		return nil, err
 	}
-	if _, err := query.SetBuffer("cols", buffD2); err != nil {
+	if _, err := query.SetDataBuffer("rows", buffD1); err != nil {
+		return nil, err
+	}
+	if _, err := query.SetDataBuffer("cols", buffD2); err != nil {
 		return nil, err
 	}
 
