@@ -3,6 +3,7 @@ package tiledb
 import (
 	"os"
 	"testing"
+	"unsafe"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -52,28 +53,28 @@ func ExampleNewQuery() {
 		return
 	}
 
-	// Crete attribute to add to schema
+	// Create attribute to add to schema
 	attribute, err := NewAttribute(context, "a1", TILEDB_INT32)
 	if err != nil {
 		// Handle error
 		return
 	}
 
-	// Crete attribute to add to schema
+	// Create attribute to add to schema
 	attribute2, err := NewAttribute(context, "a2", TILEDB_STRING_ASCII)
 	if err != nil {
 		// Handle error
 		return
 	}
 
-	// Crete attribute to add to schema
+	// Create attribute to add to schema
 	attribute3, err := NewAttribute(context, "a3", TILEDB_FLOAT32)
 	if err != nil {
 		// Handle error
 		return
 	}
 
-	// Crete attribute to add to schema
+	// Create attribute to add to schema
 	attribute4, err := NewAttribute(context, "a4", TILEDB_STRING_UTF8)
 	if err != nil {
 		// Handle error
@@ -152,8 +153,22 @@ func ExampleNewQuery() {
 		return
 	}
 
+	// Create subarray
+	subarray, err := array.NewSubarray()
+	if err != nil {
+		// Handle error
+		return
+	}
+
 	// Limit writting to subarray
-	err = query.SetSubArray([]int8{0, 1})
+	err = subarray.SetSubArray([]int8{0, 1})
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Set subarray to query
+	err = query.SetSubarray(subarray)
 	if err != nil {
 		// Handle error
 		return
@@ -168,14 +183,14 @@ func ExampleNewQuery() {
 
 	// Create write buffers
 	bufferA1 := []int32{1, 2}
-	_, err = query.SetBuffer("a1", bufferA1)
+	_, err = query.SetDataBuffer("a1", bufferA1)
 	if err != nil {
 		// Handle error
 		return
 	}
 
 	bufferA2 := []byte("ab")
-	_, err = query.SetBuffer("a2", bufferA2)
+	_, err = query.SetDataBuffer("a2", bufferA2)
 	if err != nil {
 		// Handle error
 		return
@@ -183,7 +198,12 @@ func ExampleNewQuery() {
 
 	bufferA3 := []float32{1.0, 2.0, 3.0, 4.0, 5.0}
 	offsetBufferA3 := []uint64{0, 3}
-	_, _, err = query.SetBufferVar("a3", offsetBufferA3, bufferA3)
+	_, err = query.SetDataBuffer("a3", bufferA3)
+	if err != nil {
+		// Handle error
+		return
+	}
+	_, err = query.SetOffsetsBuffer("a3", offsetBufferA3)
 	if err != nil {
 		// Handle error
 		return
@@ -191,7 +211,12 @@ func ExampleNewQuery() {
 
 	bufferA4 := []byte("hello" + "world")
 	offsetBufferA4 := []uint64{0, 5}
-	_, _, err = query.SetBufferVar("a4", offsetBufferA4, bufferA4)
+	_, err = query.SetDataBuffer("a4", bufferA4)
+	if err != nil {
+		// Handle error
+		return
+	}
+	_, err = query.SetOffsetsBuffer("a4", offsetBufferA4)
 	if err != nil {
 		// Handle error
 		return
@@ -243,8 +268,22 @@ func ExampleNewQuery() {
 		return
 	}
 
+	// Create subarray
+	subarray, err = array.NewSubarray()
+	if err != nil {
+		// Handle error
+		return
+	}
+
 	// Set read subarray to only data that was written
-	err = query.SetSubArray([]int8{0, 1})
+	err = subarray.SetSubArray([]int8{0, 1})
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Set subarray to query
+	err = query.SetSubarray(subarray)
 	if err != nil {
 		// Handle error
 		return
@@ -252,14 +291,14 @@ func ExampleNewQuery() {
 
 	// Set empty buffers for reading
 	readBufferA1 := make([]int32, 2)
-	_, err = query.SetBuffer("a1", readBufferA1)
+	_, err = query.SetDataBuffer("a1", readBufferA1)
 	if err != nil {
 		// Handle error
 		return
 	}
 
 	readBufferA2 := make([]byte, 2)
-	_, err = query.SetBuffer("a2", readBufferA2)
+	_, err = query.SetDataBuffer("a2", readBufferA2)
 	if err != nil {
 		// Handle error
 		return
@@ -267,14 +306,24 @@ func ExampleNewQuery() {
 
 	readBufferA3 := make([]float32, 5)
 	readOffsetBufferA3 := make([]uint64, 2)
-	_, _, err = query.SetBufferVar("a3", readOffsetBufferA3, readBufferA3)
+	_, err = query.SetDataBuffer("a3", readBufferA3)
+	if err != nil {
+		// Handle error
+		return
+	}
+	_, err = query.SetOffsetsBuffer("a3", readOffsetBufferA3)
 	if err != nil {
 		// Handle error
 		return
 	}
 	readBufferA4 := make([]byte, 10)
 	readOffsetBufferA4 := make([]uint64, 2)
-	_, _, err = query.SetBufferVar("a4", readOffsetBufferA4, readBufferA4)
+	_, err = query.SetDataBuffer("a4", readBufferA4)
+	if err != nil {
+		// Handle error
+		return
+	}
+	_, err = query.SetOffsetsBuffer("a4", readOffsetBufferA4)
 	if err != nil {
 		// Handle error
 		return
@@ -285,9 +334,8 @@ func ExampleNewQuery() {
 		// Handle error
 		return
 	}
-	// Submit read query async
-	// Async submits do not block
-	err = query.SubmitAsync()
+	// Submit read query
+	err = query.Submit()
 	if err != nil {
 		// Handle error
 		return
@@ -320,6 +368,271 @@ func ExampleNewQuery() {
 
 	if len(stats) > 0 {
 		// Do something with stats
+	}
+}
+
+// ExampleNewQueryCondition shows a complete write, delete and read example
+func ExampleNewQueryCondition() {
+	// Create configuration
+	config, err := NewConfig()
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Test context with config
+	context, err := NewContext(config)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Test create dimension
+	dimension, err := NewDimension(context, "dim1", TILEDB_INT8, []int8{0, 9}, int8(10))
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Test creating domain
+	domain, err := NewDomain(context)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Add dimension
+	err = domain.AddDimensions(dimension)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Create array schema
+	arraySchema, err := NewArraySchema(context, TILEDB_SPARSE)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Create attribute to add to schema
+	attribute, err := NewAttribute(context, "a1", TILEDB_INT32)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Add Attribute
+	err = arraySchema.AddAttributes(attribute)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Set Domain
+	err = arraySchema.SetDomain(domain)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	err = arraySchema.SetCellOrder(TILEDB_ROW_MAJOR)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	err = arraySchema.SetTileOrder(TILEDB_ROW_MAJOR)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Validate Schema
+	err = arraySchema.Check()
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// create temp group name
+	tmpArrayPath := os.TempDir()
+	// Create new array struct
+	array, err := NewArray(context, tmpArrayPath)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Create array on disk
+	err = array.Create(arraySchema)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Open array for writting
+	err = array.Open(TILEDB_WRITE)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Create write query
+	query, err := NewQuery(context, array)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Set write layout
+	err = query.SetLayout(TILEDB_UNORDERED)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Create write buffers
+	bufferA1 := []int32{1, 2, 3, 4}
+	_, err = query.SetDataBuffer("a1", bufferA1)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Set coordinates, since test is 1d, this is subarray
+	subArray := []int8{0, 1, 2, 3}
+	_, err = query.SetDataBuffer("dim1", subArray)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Submit write query
+	err = query.Submit()
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Validate query type
+	_, err = query.Type()
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Finalize Write
+	err = query.Finalize()
+	if err != nil {
+		// Handle error
+		return
+	}
+	// Close and prepare to delete
+	err = array.Close()
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Prepare a Delete query for elements dim1[2,3]
+	// Reopen array for deletion
+	err = array.Open(TILEDB_DELETE)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	query, err = NewQuery(context, array)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	condition, err := NewQueryCondition(context, "dim1", TILEDB_QUERY_CONDITION_GE, int8(2))
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	err = query.SetQueryCondition(condition)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// submit and finalize query
+	err = query.Submit()
+	if err != nil {
+		// Handle error
+		return
+	}
+	err = query.Finalize()
+	if err != nil {
+		// Handle error
+		return
+	}
+	// Close and prepare to read
+	err = array.Close()
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Reopen array for reading
+	err = array.Open(TILEDB_READ)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	query, err = NewQuery(context, array)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Set coordinates, since test is 1d, this is subarray
+	_, err = query.SetDataBuffer("dim1", subArray)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Set empty buffers for reading
+	readBufferA1 := make([]int32, 4)
+	_, err = query.SetDataBuffer("a1", readBufferA1)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Set read layout
+	err = query.SetLayout(TILEDB_ROW_MAJOR)
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Submit read query async
+	err = query.Submit()
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Validate query type
+	_, err = query.Type()
+	if err != nil {
+		// Handle error
+		return
+	}
+
+	// Results should be returned
+	_, err = query.HasResults()
+	if err != nil {
+		// Handle error
+		return
 	}
 }
 
@@ -409,22 +722,26 @@ func TestQueryEffectiveBufferSize(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 	require.NoError(t, query.SetLayout(TILEDB_GLOBAL_ORDER))
-	_, _, err = query.SetBufferVar("a1", a1OffWrite, a1DataWrite)
+	_, err = query.SetDataBuffer("a1", a1DataWrite)
 	require.NoError(t, err)
-	_, err = query.SetBuffer("rows", buffD1)
+	_, err = query.SetOffsetsBuffer("a1", a1OffWrite)
 	require.NoError(t, err)
-	_, err = query.SetBuffer("cols", buffD2)
+	_, err = query.SetDataBuffer("rows", buffD1)
+	require.NoError(t, err)
+	_, err = query.SetDataBuffer("cols", buffD2)
 	require.NoError(t, err)
 
 	// Check the buffer sizes
-	offsetSize, dataSize, err := query.BufferSizeVar("a1")
+	dataSize, err := query.GetExpectedDataBufferLength("a1")
+	require.NoError(t, err)
+	offsetSize, err := query.GetExpectedOffsetsBufferLength("a1")
 	require.NoError(t, err)
 	assert.Equal(t, len(a1OffWrite), int(offsetSize))
 	assert.Equal(t, len(a1DataWrite), int(dataSize))
-	rowsDataSize, err := query.BufferSize("rows")
+	rowsDataSize, err := query.GetExpectedDataBufferLength("rows")
 	require.NoError(t, err)
 	assert.Equal(t, len(buffD1), int(rowsDataSize))
-	colsDataSize, err := query.BufferSize("cols")
+	colsDataSize, err := query.GetExpectedDataBufferLength("cols")
 	require.NoError(t, err)
 	assert.Equal(t, len(buffD2), int(colsDataSize))
 
@@ -436,7 +753,11 @@ func TestQueryEffectiveBufferSize(t *testing.T) {
 	require.NoError(t, array.Open(TILEDB_READ))
 
 	// Read value at cell 2, 2
-	subArray := []int32{2, 2, 2, 2}
+	subarray, err := array.NewSubarray()
+	require.NoError(t, err)
+	assert.NotNil(t, subarray)
+	err = subarray.SetSubArray([]int32{2, 2, 2, 2})
+	require.NoError(t, err)
 
 	// Prepare buffers
 	rows := make([]int32, 2)
@@ -450,15 +771,16 @@ func TestQueryEffectiveBufferSize(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 
-	require.NoError(t, query.SetSubArray(subArray))
+	require.NoError(t, query.SetSubarray(subarray))
 	require.NoError(t, query.SetLayout(TILEDB_ROW_MAJOR))
-	offsetBufferSize, effectiveBufferSize, err := query.SetBufferVar("a1",
-		a1OffRead, a1DataRead)
+	effectiveBufferSize, err := query.SetDataBuffer("a1", a1DataRead)
+	require.NoError(t, err)
+	offsetBufferSize, err := query.SetOffsetsBuffer("a1", a1OffRead)
 	require.NoError(t, err)
 	assert.NotNil(t, query)
-	_, err = query.SetBuffer("rows", rows)
+	_, err = query.SetDataBuffer("rows", rows)
 	require.NoError(t, err)
-	_, err = query.SetBuffer("cols", cols)
+	_, err = query.SetDataBuffer("cols", cols)
 	require.NoError(t, err)
 
 	// Submit the query
@@ -580,34 +902,50 @@ func TestQueryEffectiveBufferSizeHeterogeneous(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 	require.NoError(t, query.SetLayout(TILEDB_GLOBAL_ORDER))
-	_, _, err = query.SetBufferVar("a1", a1OffWrite, a1DataWrite)
+	_, err = query.SetDataBuffer("a1", a1DataWrite)
 	require.NoError(t, err)
-	_, err = query.SetBuffer("rows", rowsWrite)
+	_, err = query.SetOffsetsBuffer("a1", a1OffWrite)
 	require.NoError(t, err)
-	_, err = query.SetBuffer("cols", colsWrite)
+	_, err = query.SetDataBuffer("rows", rowsWrite)
 	require.NoError(t, err)
-	_, _, _, err = query.SetBufferVarNullable("a2", a2OffWrite, a2DataWrite, a2Validity)
+	_, err = query.SetDataBuffer("cols", colsWrite)
 	require.NoError(t, err)
-	_, _, err = query.SetBufferNullable("a3", a3DataWrite, a3Validity)
+	_, err = query.SetDataBuffer("a2", a2DataWrite)
+	require.NoError(t, err)
+	_, err = query.SetOffsetsBuffer("a2", a2OffWrite)
+	require.NoError(t, err)
+	_, err = query.SetValidityBuffer("a2", a2Validity)
+	require.NoError(t, err)
+	_, err = query.SetDataBuffer("a3", a3DataWrite)
+	require.NoError(t, err)
+	_, err = query.SetValidityBuffer("a3", a3Validity)
 	require.NoError(t, err)
 
 	// Check the buffer sizes
-	offsetSize, dataSize, err := query.BufferSizeVar("a1")
+	dataSize, err := query.GetExpectedDataBufferLength("a1")
+	require.NoError(t, err)
+	offsetSize, err := query.GetExpectedOffsetsBufferLength("a1")
 	require.NoError(t, err)
 	assert.Equal(t, len(a1OffWrite), int(offsetSize))
 	assert.Equal(t, len(a1DataWrite), int(dataSize))
-	dataSize, err = query.BufferSize("rows")
+	dataSize, err = query.GetExpectedDataBufferLength("rows")
 	require.NoError(t, err)
 	assert.Equal(t, len(rowsWrite), int(dataSize))
-	dataSize, err = query.BufferSize("cols")
+	dataSize, err = query.GetExpectedDataBufferLength("cols")
 	require.NoError(t, err)
 	assert.Equal(t, len(colsWrite), int(dataSize))
-	offsetSize, dataSize, validitySize, err := query.BufferSizeVarNullable("a2")
+	dataSize, err = query.GetExpectedDataBufferLength("a2")
+	require.NoError(t, err)
+	offsetSize, err = query.GetExpectedOffsetsBufferLength("a2")
+	require.NoError(t, err)
+	validitySize, err := query.GetExpectedValidityBufferLength("a2")
 	require.NoError(t, err)
 	assert.Equal(t, len(a2OffWrite), int(offsetSize))
 	assert.Equal(t, len(a2DataWrite), int(dataSize))
 	assert.Equal(t, len(a2Validity), int(validitySize))
-	dataSize, validitySize, err = query.BufferSizeNullable("a3")
+	dataSize, err = query.GetExpectedDataBufferLength("a3")
+	require.NoError(t, err)
+	validitySize, err = query.GetExpectedValidityBufferLength("a3")
 	require.NoError(t, err)
 	assert.Equal(t, len(a3DataWrite), int(dataSize))
 	assert.Equal(t, len(a3Validity), int(validitySize))
@@ -644,29 +982,36 @@ func TestQueryEffectiveBufferSizeHeterogeneous(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 
-	require.NoError(t, query.AddRange(0, rowsRange[0], rowsRange[1]))
-	require.NoError(t, query.AddRange(1, colsRange[0], colsRange[1]))
+	// Prepare the subarray
+	subarray, err := array.NewSubarray()
+	require.NoError(t, err)
+	assert.NotNil(t, subarray)
+	require.NoError(t, subarray.AddRange(0, MakeRange(rowsRange[0], rowsRange[1])))
+	require.NoError(t, subarray.AddRange(1, MakeRange(colsRange[0], colsRange[1])))
+
+	require.NoError(t, query.SetSubarray(subarray))
 	require.NoError(t, query.SetLayout(TILEDB_ROW_MAJOR))
-	offsetBufferSize, effectiveBufferSize, err := query.SetBufferVar("a1",
-		a1OffRead, a1DataRead)
+	effectiveBufferSize, err := query.SetDataBuffer("a1", a1DataRead)
+	require.NoError(t, err)
+	offsetBufferSize, err := query.SetOffsetsBuffer("a1", a1OffRead)
 	require.NoError(t, err)
 	assert.NotNil(t, query)
-	_, err = query.SetBuffer("rows", rowsRead)
+	_, err = query.SetDataBuffer("rows", rowsRead)
 	require.NoError(t, err)
-	_, err = query.SetBuffer("cols", colsRead)
+	_, err = query.SetDataBuffer("cols", colsRead)
 	require.NoError(t, err)
 
 	// Get Range for rows
-	rangeStart, rangeEnd, err := query.GetRange(0, 0)
+	r, err := subarray.GetRange(0, 0)
 	require.NoError(t, err)
-	assert.EqualValues(t, rowsRange[0], rangeStart)
-	assert.EqualValues(t, rowsRange[1], rangeEnd)
+	assert.EqualValues(t, rowsRange[0], r.start)
+	assert.EqualValues(t, rowsRange[1], r.end)
 
 	// Get Range for cols
-	rangeStart, rangeEnd, err = query.GetRange(1, 0)
+	r, err = subarray.GetRange(1, 0)
 	require.NoError(t, err)
-	assert.EqualValues(t, colsRange[0], rangeStart)
-	assert.EqualValues(t, colsRange[1], rangeEnd)
+	assert.EqualValues(t, colsRange[0], r.start)
+	assert.EqualValues(t, colsRange[1], r.end)
 
 	// Submit the query
 	require.NoError(t, query.Submit())
@@ -694,34 +1039,41 @@ func TestQueryEffectiveBufferSizeHeterogeneous(t *testing.T) {
 	// Reopen the array
 	require.NoError(t, array.Open(TILEDB_READ))
 
-	// Prepare the query for add / get ranges by name
+	// Prepare the query
 	query, err = NewQuery(context, array)
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 
-	require.NoError(t, query.AddRangeByName("rows", rowsRange[0], rowsRange[1]))
-	require.NoError(t, query.AddRangeByName("cols", colsRange[0], colsRange[1]))
+	// Prepare the subarray for add / get ranges by name
+	subarray, err = array.NewSubarray()
+	require.NoError(t, err)
+	assert.NotNil(t, subarray)
+	require.NoError(t, subarray.AddRangeByName("rows", MakeRange(rowsRange[0], rowsRange[1])))
+	require.NoError(t, subarray.AddRangeByName("cols", MakeRange(colsRange[0], colsRange[1])))
+
+	require.NoError(t, query.SetSubarray(subarray))
 	require.NoError(t, query.SetLayout(TILEDB_ROW_MAJOR))
-	offsetBufferSize, effectiveBufferSize, err = query.SetBufferVar("a1",
-		a1OffRead, a1DataRead)
+	effectiveBufferSize, err = query.SetDataBuffer("a1", a1DataRead)
+	require.NoError(t, err)
+	offsetBufferSize, err = query.SetOffsetsBuffer("a1", a1OffRead)
 	require.NoError(t, err)
 	assert.NotNil(t, query)
-	_, err = query.SetBuffer("rows", rowsRead)
+	_, err = query.SetDataBuffer("rows", rowsRead)
 	require.NoError(t, err)
-	_, err = query.SetBuffer("cols", colsRead)
+	_, err = query.SetDataBuffer("cols", colsRead)
 	require.NoError(t, err)
 
 	// Get Range for rows
-	rangeStart, rangeEnd, err = query.GetRangeFromName("rows", 0)
+	r, err = subarray.GetRangeFromName("rows", 0)
 	require.NoError(t, err)
-	assert.EqualValues(t, rowsRange[0], rangeStart)
-	assert.EqualValues(t, rowsRange[1], rangeEnd)
+	assert.EqualValues(t, rowsRange[0], r.start)
+	assert.EqualValues(t, rowsRange[1], r.end)
 
 	// Get Range for cols
-	rangeStart, rangeEnd, err = query.GetRangeFromName("cols", 0)
+	r, err = subarray.GetRangeFromName("cols", 0)
 	require.NoError(t, err)
-	assert.EqualValues(t, colsRange[0], rangeStart)
-	assert.EqualValues(t, colsRange[1], rangeEnd)
+	assert.EqualValues(t, rowsRange[0], r.start)
+	assert.EqualValues(t, rowsRange[1], r.end)
 
 	// Submit the query
 	require.NoError(t, query.Submit())
@@ -816,17 +1168,25 @@ func TestQueryEffectiveBufferSizeStrings(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 	require.NoError(t, query.SetLayout(TILEDB_GLOBAL_ORDER))
-	_, _, err = query.SetBufferVar("a1", a1OffWrite, a1DataWrite)
+	_, err = query.SetDataBuffer("a1", a1DataWrite)
 	require.NoError(t, err)
-	_, _, err = query.SetBufferVar("rows", rowsOffWrite, rowsWrite)
+	_, err = query.SetOffsetsBuffer("a1", a1OffWrite)
+	require.NoError(t, err)
+	_, err = query.SetDataBuffer("rows", rowsWrite)
+	require.NoError(t, err)
+	_, err = query.SetOffsetsBuffer("rows", rowsOffWrite)
 	require.NoError(t, err)
 
 	// Check the buffer sizes
-	offsetSize, dataSize, err := query.BufferSizeVar("a1")
+	dataSize, err := query.GetExpectedDataBufferLength("a1")
+	require.NoError(t, err)
+	offsetSize, err := query.GetExpectedOffsetsBufferLength("a1")
 	require.NoError(t, err)
 	assert.Equal(t, len(a1OffWrite), int(offsetSize))
 	assert.Equal(t, len(a1DataWrite), int(dataSize))
-	offsetSize, dataSize, err = query.BufferSizeVar("rows")
+	dataSize, err = query.GetExpectedDataBufferLength("rows")
+	require.NoError(t, err)
+	offsetSize, err = query.GetExpectedOffsetsBufferLength("rows")
 	require.NoError(t, err)
 	assert.Equal(t, len(rowsOffWrite), int(offsetSize))
 	assert.Equal(t, len(rowsWrite), int(dataSize))
@@ -839,7 +1199,7 @@ func TestQueryEffectiveBufferSizeStrings(t *testing.T) {
 	require.NoError(t, array.Open(TILEDB_READ))
 
 	// Read value at cell "bb"
-	rowsRange := [][]byte{[]byte("bb"), []byte("bb")}
+	rowsRange := []string{"bb", "bb"}
 
 	// Prepare buffers
 	rowsRead := make([]byte, 4)
@@ -853,20 +1213,28 @@ func TestQueryEffectiveBufferSizeStrings(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 
-	require.NoError(t, query.AddRangeVar(0, rowsRange[0], rowsRange[1]))
+	subarray, err := array.NewSubarray()
+	require.NoError(t, err)
+	assert.NotNil(t, subarray)
+	require.NoError(t, subarray.AddRange(0, MakeRange(rowsRange[0], rowsRange[1])))
+
+	require.NoError(t, query.SetSubarray(subarray))
 	require.NoError(t, query.SetLayout(TILEDB_ROW_MAJOR))
-	offsetBufferSize, effectiveBufferSize, err := query.SetBufferVar("a1",
-		a1OffRead, a1DataRead)
+	effectiveBufferSize, err := query.SetDataBuffer("a1", a1DataRead)
+	require.NoError(t, err)
+	offsetBufferSize, err := query.SetOffsetsBuffer("a1", a1OffRead)
 	require.NoError(t, err)
 	assert.NotNil(t, query)
-	_, _, err = query.SetBufferVar("rows", rowsOffRead, rowsRead)
+	_, err = query.SetDataBuffer("rows", rowsRead)
+	require.NoError(t, err)
+	_, err = query.SetOffsetsBuffer("rows", rowsOffRead)
 	require.NoError(t, err)
 
 	// Get Range
-	rangeStart, rangeEnd, err := query.GetRange(0, 0)
+	r, err := subarray.GetRange(0, 0)
 	require.NoError(t, err)
-	assert.EqualValues(t, rowsRange[0], rangeStart)
-	assert.EqualValues(t, rowsRange[1], rangeEnd)
+	assert.EqualValues(t, rowsRange[0], r.start)
+	assert.EqualValues(t, rowsRange[1], r.end)
 
 	// Submit the query
 	require.NoError(t, query.Submit())
@@ -905,20 +1273,28 @@ func TestQueryEffectiveBufferSizeStrings(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 
-	require.NoError(t, query.AddRangeVarByName("rows", rowsRange[0], rowsRange[1]))
+	subarray, err = array.NewSubarray()
+	require.NoError(t, err)
+	assert.NotNil(t, subarray)
+	require.NoError(t, subarray.AddRangeByName("rows", MakeRange(rowsRange[0], rowsRange[1])))
+
+	require.NoError(t, query.SetSubarray(subarray))
 	require.NoError(t, query.SetLayout(TILEDB_ROW_MAJOR))
-	offsetBufferSize, effectiveBufferSize, err = query.SetBufferVar("a1",
-		a1OffRead, a1DataRead)
+	effectiveBufferSize, err = query.SetDataBuffer("a1", a1DataRead)
+	require.NoError(t, err)
+	offsetBufferSize, err = query.SetOffsetsBuffer("a1", a1OffRead)
 	require.NoError(t, err)
 	assert.NotNil(t, query)
-	_, _, err = query.SetBufferVar("rows", rowsOffRead, rowsRead)
+	_, err = query.SetDataBuffer("rows", rowsRead)
+	require.NoError(t, err)
+	_, err = query.SetOffsetsBuffer("rows", rowsOffRead)
 	require.NoError(t, err)
 
 	// Get Range
-	rangeStart, rangeEnd, err = query.GetRangeFromName("rows", 0)
+	r, err = subarray.GetRangeFromName("rows", 0)
 	require.NoError(t, err)
-	assert.EqualValues(t, rowsRange[0], rangeStart)
-	assert.EqualValues(t, rowsRange[1], rangeEnd)
+	assert.EqualValues(t, rowsRange[0], r.start)
+	assert.EqualValues(t, rowsRange[1], r.end)
 
 	// Submit the query
 	require.NoError(t, query.Submit())
@@ -1016,23 +1392,31 @@ func TestQueryEffectiveBufferSizeStringsHeterogeneous(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 	require.NoError(t, query.SetLayout(TILEDB_GLOBAL_ORDER))
-	_, _, err = query.SetBufferVar("a1", a1OffWrite, a1DataWrite)
+	_, err = query.SetDataBuffer("a1", a1DataWrite)
 	require.NoError(t, err)
-	_, _, err = query.SetBufferVar("rows", rowsOffWrite, rowsWrite)
+	_, err = query.SetOffsetsBuffer("a1", a1OffWrite)
 	require.NoError(t, err)
-	_, err = query.SetBuffer("cols", colsWrite)
+	_, err = query.SetDataBuffer("rows", rowsWrite)
+	require.NoError(t, err)
+	_, err = query.SetOffsetsBuffer("rows", rowsOffWrite)
+	require.NoError(t, err)
+	_, err = query.SetDataBuffer("cols", colsWrite)
 	require.NoError(t, err)
 
 	// Check the buffer sizes
-	offsetSize, dataSize, err := query.BufferSizeVar("a1")
+	dataSize, err := query.GetExpectedDataBufferLength("a1")
+	require.NoError(t, err)
+	offsetSize, err := query.GetExpectedOffsetsBufferLength("a1")
 	require.NoError(t, err)
 	assert.Equal(t, len(a1OffWrite), int(offsetSize))
 	assert.Equal(t, len(a1DataWrite), int(dataSize))
-	offsetSize, dataSize, err = query.BufferSizeVar("rows")
+	dataSize, err = query.GetExpectedDataBufferLength("rows")
+	require.NoError(t, err)
+	offsetSize, err = query.GetExpectedOffsetsBufferLength("rows")
 	require.NoError(t, err)
 	assert.Equal(t, len(rowsOffWrite), int(offsetSize))
 	assert.Equal(t, len(rowsWrite), int(dataSize))
-	dataSize, err = query.BufferSize("cols")
+	dataSize, err = query.GetExpectedDataBufferLength("cols")
 	require.NoError(t, err)
 	assert.Equal(t, len(colsWrite), int(dataSize))
 
@@ -1044,7 +1428,7 @@ func TestQueryEffectiveBufferSizeStringsHeterogeneous(t *testing.T) {
 	require.NoError(t, array.Open(TILEDB_READ))
 
 	// Read value at cell "c", 2
-	rowsRange := [][]byte{[]byte("c"), []byte("c")}
+	rowsRange := []string{"c", "c"}
 	colsRange := []int64{2, 2}
 
 	// Prepare buffers
@@ -1060,16 +1444,25 @@ func TestQueryEffectiveBufferSizeStringsHeterogeneous(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 
-	require.NoError(t, query.AddRangeVar(0, rowsRange[0], rowsRange[1]))
-	require.NoError(t, query.AddRange(1, colsRange[0], colsRange[1]))
+	// Prepare the subarray
+	subarray, err := array.NewSubarray()
+	require.NoError(t, err)
+	assert.NotNil(t, subarray)
+
+	require.NoError(t, subarray.AddRange(0, MakeRange(rowsRange[0], rowsRange[1])))
+	require.NoError(t, subarray.AddRange(1, MakeRange(colsRange[0], colsRange[1])))
+	require.NoError(t, query.SetSubarray(subarray))
 	require.NoError(t, query.SetLayout(TILEDB_ROW_MAJOR))
-	offsetBufferSize, effectiveBufferSize, err := query.SetBufferVar("a1",
-		a1OffRead, a1DataRead)
+	effectiveBufferSize, err := query.SetDataBuffer("a1", a1DataRead)
+	require.NoError(t, err)
+	offsetBufferSize, err := query.SetOffsetsBuffer("a1", a1OffRead)
 	require.NoError(t, err)
 	assert.NotNil(t, query)
-	_, _, err = query.SetBufferVar("rows", rowsOffRead, rowsRead)
+	_, err = query.SetDataBuffer("rows", rowsRead)
 	require.NoError(t, err)
-	_, err = query.SetBuffer("cols", colsRead)
+	_, err = query.SetOffsetsBuffer("rows", rowsOffRead)
+	require.NoError(t, err)
+	_, err = query.SetDataBuffer("cols", colsRead)
 	require.NoError(t, err)
 
 	// Submit the query
@@ -1138,7 +1531,7 @@ func TestQueryReadEmpty(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, attribute3)
 
-	// Crete attribute to add to schema
+	// Create attribute to add to schema
 	attribute4, err := NewAttribute(context, "a4", TILEDB_STRING_UTF8)
 	require.NoError(t, err)
 	assert.NotNil(t, attribute4)
@@ -1173,31 +1566,41 @@ func TestQueryReadEmpty(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 
+	// Create Subarray
+	subarray, err := array.NewSubarray()
+	require.NoError(t, err)
+	assert.NotNil(t, subarray)
+
 	// Limit reading to subArray
-	require.NoError(t, query.SetSubArray([]int8{2, 4}))
+	require.NoError(t, subarray.SetSubArray([]int8{2, 4}))
+	require.NoError(t, query.SetSubarray(subarray))
 
 	// Set buffer to incorrect type, should err
 	bufferA1Bad := make([]int8, 4)
-	_, err = query.SetBuffer("a1", bufferA1Bad)
+	_, err = query.SetDataBuffer("a1", bufferA1Bad)
 	assert.Error(t, err)
 
 	// Create read buffers
 	bufferA1 := make([]int32, 4)
-	_, err = query.SetBuffer("a1", bufferA1)
+	_, err = query.SetDataBuffer("a1", bufferA1)
 	require.NoError(t, err)
 
 	bufferA2 := make([]byte, 4)
-	_, err = query.SetBuffer("a2", bufferA2)
+	_, err = query.SetDataBuffer("a2", bufferA2)
 	require.NoError(t, err)
 
 	bufferA3 := make([]float32, 10)
 	offsetBufferA3 := make([]uint64, 6)
-	_, _, err = query.SetBufferVar("a3", offsetBufferA3, bufferA3)
+	_, err = query.SetDataBuffer("a3", bufferA3)
+	require.NoError(t, err)
+	_, err = query.SetOffsetsBuffer("a3", offsetBufferA3)
 	require.NoError(t, err)
 
 	bufferA4 := make([]byte, 8)
 	offsetBufferA4 := make([]uint64, 8)
-	_, _, err = query.SetBufferVar("a4", offsetBufferA4, bufferA4)
+	_, err = query.SetDataBuffer("a4", bufferA4)
+	require.NoError(t, err)
+	_, err = query.SetOffsetsBuffer("a4", offsetBufferA4)
 	require.NoError(t, err)
 
 	// Set read layout
@@ -1251,32 +1654,32 @@ func TestDenseQueryWrite(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, arraySchema)
 
-	// Crete attribute to add to schema
+	// Create attribute to add to schema
 	attribute, err := NewAttribute(context, "a1", TILEDB_INT32)
 	require.NoError(t, err)
 	assert.NotNil(t, attribute)
 
-	// Crete attribute to add to schema
+	// Create attribute to add to schema
 	attribute2, err := NewAttribute(context, "a2", TILEDB_STRING_ASCII)
 	require.NoError(t, err)
 	assert.NotNil(t, attribute2)
 
-	// Crete attribute to add to schema
+	// Create attribute to add to schema
 	attribute3, err := NewAttribute(context, "a3", TILEDB_FLOAT32)
 	require.NoError(t, err)
 	assert.NotNil(t, attribute3)
 
-	// Crete attribute to add to schema
+	// Create attribute to add to schema
 	attribute4, err := NewAttribute(context, "a4", TILEDB_STRING_UTF8)
 	require.NoError(t, err)
 	assert.NotNil(t, attribute4)
 
-	// Crete attribute to add to schema
+	// Create attribute to add to schema
 	attribute5, err := NewAttribute(context, "a5", TILEDB_CHAR)
 	require.NoError(t, err)
 	assert.NotNil(t, attribute5)
 
-	// Crete attribute to add to schema
+	// Create attribute to add to schema
 	attribute6, err := NewAttribute(context, "a6", TILEDB_CHAR)
 	require.NoError(t, err)
 	assert.NotNil(t, attribute5)
@@ -1328,22 +1731,29 @@ func TestDenseQueryWrite(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 
-	require.NoError(t, query.SetSubArray([]int8{0, 1}))
+	// Create subarray
+	subarray, err := array.NewSubarray()
+	require.NoError(t, err)
+	assert.NotNil(t, subarray)
+	require.NoError(t, subarray.SetSubArray([]int8{0, 1}))
+	require.NoError(t, query.SetSubarray(subarray))
 
 	// Set write layout
 	assert.Nil(t, query.SetLayout(TILEDB_ROW_MAJOR))
 
 	bufferA1 := []int32{1, 2}
-	_, err = query.SetBuffer("a1", bufferA1)
+	_, err = query.SetDataBuffer("a1", bufferA1)
 	require.NoError(t, err)
 
 	bufferA2 := []byte("ab")
-	_, err = query.SetBuffer("a2", bufferA2)
+	_, err = query.SetDataBuffer("a2", bufferA2)
 	require.NoError(t, err)
 
 	bufferA3 := []float32{1.0, 2.0, 3.0, 4.0, 5.0}
 	offsetBufferA3 := []uint64{0, 3}
-	_, _, err = query.SetBufferVar("a3", offsetBufferA3, bufferA3)
+	_, err = query.SetDataBuffer("a3", bufferA3)
+	require.NoError(t, err)
+	_, err = query.SetOffsetsBuffer("a3", offsetBufferA3)
 	require.NoError(t, err)
 
 	bufferA4 := []byte("hello" + "world")
@@ -1353,7 +1763,9 @@ func TestDenseQueryWrite(t *testing.T) {
 	elementsCopied := copy(bufferA4Comparison, bufferA4)
 	assert.Equal(t, len(bufferA4), elementsCopied)
 
-	_, _, err = query.SetBufferVar("a4", offsetBufferA4, bufferA4)
+	_, err = query.SetDataBuffer("a4", bufferA4)
+	require.NoError(t, err)
+	_, err = query.SetOffsetsBuffer("a4", offsetBufferA4)
 	require.NoError(t, err)
 
 	bufferA5 := "hello" + "world"
@@ -1365,12 +1777,16 @@ func TestDenseQueryWrite(t *testing.T) {
 	assert.EqualValues(t, bufferA5, bufferA5Comparison)
 	bufferA5Bytes := []byte(bufferA5)
 
-	_, _, err = query.SetBufferVar("a5", offsetBufferA5, bufferA5Bytes)
+	_, err = query.SetDataBuffer("a5", bufferA5Bytes)
+	require.NoError(t, err)
+	_, err = query.SetOffsetsBuffer("a5", offsetBufferA5)
 	require.NoError(t, err)
 
 	bufferA6 := []byte("ab")
 	validityBufferA6 := []uint8{0, 1}
-	_, _, err = query.SetBufferNullable("a6", bufferA6, validityBufferA6)
+	_, err = query.SetDataBuffer("a6", bufferA6)
+	require.NoError(t, err)
+	_, err = query.SetValidityBuffer("a6", validityBufferA6)
 	require.NoError(t, err)
 
 	bufferA6Comparison := make([]byte, len(bufferA6))
@@ -1387,7 +1803,11 @@ func TestDenseQueryWrite(t *testing.T) {
 	offsetBufferA7 := []uint64{0, 5}
 	validityBufferA7 := []uint8{0, 1}
 	bufferA7Bytes := []byte(bufferA7)
-	_, _, _, err = query.SetBufferVarNullable("a7", offsetBufferA7, bufferA7Bytes, validityBufferA7)
+	_, err = query.SetDataBuffer("a7", bufferA7Bytes)
+	require.NoError(t, err)
+	_, err = query.SetOffsetsBuffer("a7", offsetBufferA7)
+	require.NoError(t, err)
+	_, err = query.SetValidityBuffer("a7", validityBufferA7)
 	require.NoError(t, err)
 
 	bufferA7Comparison := make([]byte, len(bufferA7))
@@ -1434,8 +1854,12 @@ func TestDenseQueryWrite(t *testing.T) {
 	assert.NotNil(t, query)
 
 	// Set read subarray to only data that was written
+	subarray, err = array.NewSubarray()
+	require.NoError(t, err)
+	assert.NotNil(t, subarray)
 	subArray := []int8{0, 1}
-	require.NoError(t, query.SetSubArray(subArray))
+	require.NoError(t, subarray.SetSubArray(subArray))
+	require.NoError(t, query.SetSubarray(subarray))
 
 	bufferElements, err := query.EstimateBufferElements()
 	require.NoError(t, err)
@@ -1452,51 +1876,60 @@ func TestDenseQueryWrite(t *testing.T) {
 
 	// Set empty buffers for reading
 	readBufferA1 := make([]int32, 2)
-	_, err = query.SetBuffer("a1", readBufferA1)
+	_, err = query.SetDataBuffer("a1", readBufferA1)
 	require.NoError(t, err)
 
 	readBufferA2 := make([]byte, 2)
-	_, err = query.SetBuffer("a2", readBufferA2)
+	_, err = query.SetDataBuffer("a2", readBufferA2)
 	require.NoError(t, err)
 
 	readBufferA3 := make([]float32, 5)
 	readOffsetBufferA3 := make([]uint64, 2)
-	_, _, err = query.SetBufferVar("a3", readOffsetBufferA3, readBufferA3)
+	_, err = query.SetDataBuffer("a3", readBufferA3)
+	require.NoError(t, err)
+	_, err = query.SetOffsetsBuffer("a3", readOffsetBufferA3)
 	require.NoError(t, err)
 
 	readBufferA4 := make([]byte, 10)
 	readOffsetBufferA4 := make([]uint64, 2)
-	_, _, err = query.SetBufferVar("a4", readOffsetBufferA4, readBufferA4)
+	_, err = query.SetDataBuffer("a4", readBufferA4)
+	require.NoError(t, err)
+	_, err = query.SetOffsetsBuffer("a4", readOffsetBufferA4)
 	require.NoError(t, err)
 
 	readBufferA5 := make([]byte, 10) // make(string, 10)
 	readOffsetBufferA5 := make([]uint64, 2)
-	_, _, err = query.SetBufferVar("a5", readOffsetBufferA5, readBufferA5)
+	_, err = query.SetDataBuffer("a5", readBufferA5)
+	require.NoError(t, err)
+	_, err = query.SetOffsetsBuffer("a5", readOffsetBufferA5)
 	require.NoError(t, err)
 
 	readBufferA6 := make([]byte, 2)
 	readValidityBufferA6 := make([]uint8, 2)
-	_, _, err = query.SetBufferNullable("a6", readBufferA6, readValidityBufferA6)
+	_, err = query.SetDataBuffer("a6", readBufferA6)
+	require.NoError(t, err)
+	_, err = query.SetValidityBuffer("a6", readValidityBufferA6)
 	require.NoError(t, err)
 
 	readBufferA7 := make([]byte, 10)
 	readOffsetBufferA7 := make([]uint64, 2)
 	readValidityBufferA7 := make([]uint8, 2)
-	_, _, _, err = query.SetBufferVarNullable("a7", readOffsetBufferA7, readBufferA7, readValidityBufferA7)
+	_, err = query.SetDataBuffer("a7", readBufferA7)
+	require.NoError(t, err)
+	_, err = query.SetOffsetsBuffer("a7", readOffsetBufferA7)
+	require.NoError(t, err)
+	_, err = query.SetValidityBuffer("a7", readValidityBufferA7)
 	require.NoError(t, err)
 
 	// Set read layout
 	require.NoError(t, query.SetLayout(TILEDB_ROW_MAJOR))
 
 	// Submit read query async
-	require.NoError(t, query.SubmitAsync())
+	require.NoError(t, query.Submit())
 
 	// Wait for status to return complete or to error
 	// Loop while status is inprogress
-	for status, err = query.Status(); status == TILEDB_INPROGRESS && err == nil; status, err = query.Status() {
-		require.NoError(t, err)
-		assert.Equal(t, TILEDB_INPROGRESS, status)
-	}
+	status, err = query.Status()
 	require.NoError(t, err)
 	assert.Equal(t, TILEDB_COMPLETED, status)
 
@@ -1519,32 +1952,198 @@ func TestDenseQueryWrite(t *testing.T) {
 	assert.EqualValues(t, bufferA6Comparison, readBufferA6)
 	assert.EqualValues(t, bufferA7Comparison, readBufferA7)
 
-	bufferA1InterfaceGet, err := query.Buffer("a1")
+	bufferA1InterfaceGet, err := query.GetDataBuffer("a1")
 	require.NoError(t, err)
 	assert.EqualValues(t, bufferA1, bufferA1InterfaceGet.([]int32))
 
-	offsetsBufferA4Get, bufferA4InterfaceGet, err := query.BufferVar("a4")
+	bufferA4InterfaceGet, err := query.GetDataBuffer("a4")
+	require.NoError(t, err)
+	offsetsBufferA4Get, err := query.GetOffsetsBuffer("a4")
 	require.NoError(t, err)
 	assert.EqualValues(t, bufferA4Comparison, bufferA4InterfaceGet.([]byte))
 	assert.EqualValues(t, offsetBufferA4, offsetsBufferA4Get)
 
-	offsetsBufferA5Get, bufferA5InterfaceGet, err := query.BufferVar("a5")
+	bufferA5InterfaceGet, err := query.GetDataBuffer("a5")
+	require.NoError(t, err)
+	offsetsBufferA5Get, err := query.GetOffsetsBuffer("a5")
 	require.NoError(t, err)
 	assert.EqualValues(t, bufferA5Comparison, bufferA5InterfaceGet.([]byte))
 	assert.EqualValues(t, offsetBufferA5, offsetsBufferA5Get)
 
-	bufferA6InterfaceGet, bufferA6ValidityGet, err := query.BufferNullable("a6")
+	bufferA6InterfaceGet, err := query.GetDataBuffer("a6")
+	require.NoError(t, err)
+	bufferA6ValidityGet, err := query.GetValidityBuffer("a6")
 	require.NoError(t, err)
 	assert.EqualValues(t, bufferA6Comparison, bufferA6InterfaceGet.([]byte))
 	assert.EqualValues(t, bufferA6ValidityComparison, bufferA6ValidityGet)
 
-	offsetsBufferA7Get, bufferA7InterfaceGet, bufferA7ValidityGet, err := query.BufferVarNullable("a7")
+	bufferA7InterfaceGet, err := query.GetDataBuffer("a7")
+	require.NoError(t, err)
+	offsetsBufferA7Get, err := query.GetOffsetsBuffer("a7")
+	require.NoError(t, err)
+	bufferA7ValidityGet, err := query.GetValidityBuffer("a7")
 	require.NoError(t, err)
 	assert.EqualValues(t, bufferA7Comparison, bufferA7InterfaceGet.([]byte))
 	assert.EqualValues(t, offsetBufferA7, offsetsBufferA7Get)
 	assert.EqualValues(t, bufferA7ValidityComparison, bufferA7ValidityGet)
 
 	query.Free()
+}
+
+// TestSparseQueryDelete validates that sparse array elements can be deleted
+func TestSparseQueryDelete(t *testing.T) {
+	// Create configuration
+	config, err := NewConfig()
+	require.NoError(t, err)
+
+	// Test context with config
+	context, err := NewContext(config)
+	require.NoError(t, err)
+
+	// Test create dimension
+	dimension, err := NewDimension(context, "dim1", TILEDB_INT8, []int8{0, 9}, int8(10))
+	require.NoError(t, err)
+	assert.NotNil(t, dimension)
+
+	// Test creating domain
+	domain, err := NewDomain(context)
+	require.NoError(t, err)
+	assert.NotNil(t, domain)
+
+	// Add dimension
+	require.NoError(t, domain.AddDimensions(dimension))
+
+	// Create array schema
+	arraySchema, err := NewArraySchema(context, TILEDB_SPARSE)
+	require.NoError(t, err)
+	assert.NotNil(t, arraySchema)
+
+	// Create attribute to add to schema
+	attribute, err := NewAttribute(context, "a1", TILEDB_INT32)
+	require.NoError(t, err)
+	assert.NotNil(t, attribute)
+
+	// Add Attribute
+	require.NoError(t, arraySchema.AddAttributes(attribute))
+
+	// Set Domain
+	require.NoError(t, arraySchema.SetDomain(domain))
+
+	require.NoError(t, arraySchema.SetCellOrder(TILEDB_ROW_MAJOR))
+
+	require.NoError(t, arraySchema.SetTileOrder(TILEDB_ROW_MAJOR))
+
+	// Validate Schema
+	require.NoError(t, arraySchema.Check())
+
+	// create temp group name
+	tmpArrayPath := t.TempDir()
+	// Create new array struct
+	array, err := NewArray(context, tmpArrayPath)
+	require.NoError(t, err)
+	assert.NotNil(t, array)
+
+	// Create array on disk
+	require.NoError(t, array.Create(arraySchema))
+
+	// Open array for writting
+	require.NoError(t, array.Open(TILEDB_WRITE))
+
+	// Create write query
+	query, err := NewQuery(context, array)
+	require.NoError(t, err)
+	assert.NotNil(t, query)
+
+	// Set write layout
+	assert.Nil(t, query.SetLayout(TILEDB_UNORDERED))
+
+	// Create write buffers
+	bufferA1 := []int32{1, 2, 3, 4}
+	_, err = query.SetDataBuffer("a1", bufferA1)
+	require.NoError(t, err)
+
+	// Set coordinates, since test is 1d, this is subarray
+	subArray := []int8{0, 1, 2, 3}
+	_, err = query.SetDataBuffer("dim1", subArray)
+	require.NoError(t, err)
+
+	// Submit write query
+	require.NoError(t, query.Submit())
+
+	// Validate status, since query was used this is should be complete
+	status, err := query.Status()
+	require.NoError(t, err)
+	assert.Equal(t, TILEDB_COMPLETED, status)
+
+	// Validate query type
+	queryType, err := query.Type()
+	require.NoError(t, err)
+	assert.Equal(t, TILEDB_WRITE, queryType)
+
+	// Finalize Write
+	require.NoError(t, query.Finalize())
+	// Close and prepare to delete
+	require.NoError(t, array.Close())
+
+	// Prepare a Delete query for elements dim1[2,3]
+	// Reopen array for deletion
+	require.NoError(t, array.Open(TILEDB_DELETE))
+
+	query, err = NewQuery(context, array)
+	require.NoError(t, err)
+	assert.NotNil(t, query)
+
+	condition, err := NewQueryCondition(context, "dim1", TILEDB_QUERY_CONDITION_GE, int8(2))
+	require.NoError(t, err)
+
+	err = query.SetQueryCondition(condition)
+	require.NoError(t, err)
+
+	// submit and finalize query
+	require.NoError(t, query.Submit())
+	require.NoError(t, query.Finalize())
+	// Close and prepare to read
+	require.NoError(t, array.Close())
+
+	// Reopen array for reading
+	require.NoError(t, array.Open(TILEDB_READ))
+
+	query, err = NewQuery(context, array)
+	require.NoError(t, err)
+	assert.NotNil(t, query)
+
+	// Set coordinates, since test is 1d, this is subarray
+	_, err = query.SetDataBuffer("dim1", subArray)
+	require.NoError(t, err)
+
+	// Set empty buffers for reading
+	readBufferA1 := make([]int32, 4)
+	_, err = query.SetDataBuffer("a1", readBufferA1)
+	require.NoError(t, err)
+
+	// Set read layout
+	require.NoError(t, query.SetLayout(TILEDB_ROW_MAJOR))
+
+	// Submit read query
+	require.NoError(t, query.Submit())
+
+	// Get query status
+	status, err = query.Status()
+	require.NoError(t, err)
+	assert.Equal(t, TILEDB_COMPLETED, status)
+
+	// Validate query type
+	queryType, err = query.Type()
+	require.NoError(t, err)
+	assert.Equal(t, TILEDB_READ, queryType)
+
+	// Results should be returned
+	hasResults, err := query.HasResults()
+	require.NoError(t, err)
+	assert.Equal(t, true, hasResults)
+
+	// Validate read buffers equal original write buffers
+	assert.ElementsMatch(t, []int32{1, 2, 0, 0}, readBufferA1)
 }
 
 // TestSparseQueryWrite validates a sparse array can be written to and read from
@@ -1576,7 +2175,7 @@ func TestSparseQueryWrite(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, arraySchema)
 
-	// Crete attribute to add to schema
+	// Create attribute to add to schema
 	attribute, err := NewAttribute(context, "a1", TILEDB_INT32)
 	require.NoError(t, err)
 	assert.NotNil(t, attribute)
@@ -1617,12 +2216,12 @@ func TestSparseQueryWrite(t *testing.T) {
 
 	// Create write buffers
 	bufferA1 := []int32{1, 2}
-	_, err = query.SetBuffer("a1", bufferA1)
+	_, err = query.SetDataBuffer("a1", bufferA1)
 	require.NoError(t, err)
 
 	// Set coordinates, since test is 1d, this is subarray
 	subArray := []int8{0, 1}
-	_, err = query.SetBuffer("dim1", subArray)
+	_, err = query.SetDataBuffer("dim1", subArray)
 	require.NoError(t, err)
 
 	// Submit write query
@@ -1655,7 +2254,7 @@ func TestSparseQueryWrite(t *testing.T) {
 	// require.NoError(t, err)
 
 	// Set coordinates, since test is 1d, this is subarray
-	_, err = query.SetBuffer("dim1", subArray)
+	_, err = query.SetDataBuffer("dim1", subArray)
 	require.NoError(t, err)
 
 	bufferElements, err := query.EstimateBufferElements()
@@ -1665,21 +2264,17 @@ func TestSparseQueryWrite(t *testing.T) {
 
 	// Set empty buffers for reading
 	readBufferA1 := make([]int32, 2)
-	_, err = query.SetBuffer("a1", readBufferA1)
+	_, err = query.SetDataBuffer("a1", readBufferA1)
 	require.NoError(t, err)
 
 	// Set read layout
 	require.NoError(t, query.SetLayout(TILEDB_ROW_MAJOR))
 
-	// Submit read query async
-	require.NoError(t, query.SubmitAsync())
+	// Submit read query
+	require.NoError(t, query.Submit())
 
-	// Wait for status to return complete or to error
-	// Loop while status is inprogress
-	for status, err = query.Status(); status == TILEDB_INPROGRESS && err == nil; status, err = query.Status() {
-		require.NoError(t, err)
-		assert.Equal(t, TILEDB_INPROGRESS, status)
-	}
+	// Validate status, since query was used this is should be complete
+	status, err = query.Status()
 	require.NoError(t, err)
 	assert.Equal(t, TILEDB_COMPLETED, status)
 
@@ -1726,7 +2321,7 @@ func TestSparseQueryWriteNullable(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, arraySchema)
 
-	// Crete attribute to add to schema
+	// Create attribute to add to schema
 	attribute, err := NewAttribute(context, "a1", TILEDB_INT32)
 	require.NoError(t, err)
 	assert.NotNil(t, attribute)
@@ -1771,12 +2366,14 @@ func TestSparseQueryWriteNullable(t *testing.T) {
 	// Create write buffers
 	bufferA1 := []int32{1, 2, 3}
 	bufferA1Validity := []uint8{1, 1, 0}
-	_, _, err = query.SetBufferNullable("a1", bufferA1, bufferA1Validity)
+	_, err = query.SetDataBuffer("a1", bufferA1)
+	require.NoError(t, err)
+	_, err = query.SetValidityBuffer("a1", bufferA1Validity)
 	require.NoError(t, err)
 
 	// Set coordinates, since test is 1d, this is subarray
 	subArray := []int8{0, 1, 2}
-	_, err = query.SetBuffer("dim1", subArray)
+	_, err = query.SetDataBuffer("dim1", subArray)
 	require.NoError(t, err)
 
 	// Submit write query
@@ -1804,11 +2401,16 @@ func TestSparseQueryWriteNullable(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 
+	subarray, err := array.NewSubarray()
+	require.NoError(t, err)
+	assert.NotNil(t, subarray)
+
 	// Set read subarray to only data that was written
-	require.NoError(t, query.AddRange(0, 0, 3))
+	require.NoError(t, subarray.AddRange(0, MakeRange[int8](0, 3)))
+	require.NoError(t, query.SetSubarray(subarray))
 
 	// Set coordinates, since test is 1d, this is subarray
-	_, err = query.SetBuffer("dim1", subArray)
+	_, err = query.SetDataBuffer("dim1", subArray)
 	require.NoError(t, err)
 
 	bufferElements, err := query.EstimateBufferElements()
@@ -1820,7 +2422,9 @@ func TestSparseQueryWriteNullable(t *testing.T) {
 	// Set empty buffers for reading
 	readBufferA1 := make([]int32, 3)
 	readBufferA1Validity := make([]uint8, 3)
-	_, _, err = query.SetBufferNullable("a1", readBufferA1, readBufferA1Validity)
+	_, err = query.SetDataBuffer("a1", readBufferA1)
+	require.NoError(t, err)
+	_, err = query.SetValidityBuffer("a1", readBufferA1Validity)
 	require.NoError(t, err)
 
 	// Set read layout
@@ -1886,10 +2490,10 @@ func TestSparseQueryWriteHilbertLayout(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 	bufferA1 := []int32{1, 2}
-	_, err = query.SetBuffer("a1", bufferA1)
+	_, err = query.SetDataBuffer("a1", bufferA1)
 	require.NoError(t, err)
 	subArray := []int8{0, 1}
-	_, err = query.SetBuffer("dim1", subArray)
+	_, err = query.SetDataBuffer("dim1", subArray)
 	require.NoError(t, err)
 	// Set write layout
 	// Hilbert order not applicable to write queries
@@ -1903,10 +2507,10 @@ func TestSparseQueryWriteHilbertLayout(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 	bufferA1 = make([]int32, 2)
-	_, err = query.SetBuffer("a1", bufferA1)
+	_, err = query.SetDataBuffer("a1", bufferA1)
 	require.NoError(t, err)
 	subArray = make([]int8, 2)
-	_, err = query.SetBuffer("dim1", subArray)
+	_, err = query.SetDataBuffer("dim1", subArray)
 	require.NoError(t, err)
 	// Set write layout
 	// Hilbert order not applicable to write queries
@@ -2000,22 +2604,26 @@ func TestQueryConfig(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 	require.NoError(t, query.SetLayout(TILEDB_GLOBAL_ORDER))
-	_, _, err = query.SetBufferVar("a1", a1OffWrite, a1DataWrite)
+	_, err = query.SetDataBuffer("a1", a1DataWrite)
 	require.NoError(t, err)
-	_, err = query.SetBuffer("rows", buffD1)
+	_, err = query.SetOffsetsBuffer("a1", a1OffWrite)
 	require.NoError(t, err)
-	_, err = query.SetBuffer("cols", buffD2)
+	_, err = query.SetDataBuffer("rows", buffD1)
+	require.NoError(t, err)
+	_, err = query.SetDataBuffer("cols", buffD2)
 	require.NoError(t, err)
 
 	// Check the buffer sizes
-	offsetSize, dataSize, err := query.BufferSizeVar("a1")
+	dataSize, err := query.GetExpectedDataBufferLength("a1")
+	require.NoError(t, err)
+	offsetSize, err := query.GetExpectedOffsetsBufferLength("a1")
 	require.NoError(t, err)
 	assert.Equal(t, len(a1OffWrite), int(offsetSize))
 	assert.Equal(t, len(a1DataWrite), int(dataSize))
-	rowsDataSize, err := query.BufferSize("rows")
+	rowsDataSize, err := query.GetExpectedDataBufferLength("rows")
 	require.NoError(t, err)
 	assert.Equal(t, len(buffD1), int(rowsDataSize))
-	colsDataSize, err := query.BufferSize("cols")
+	colsDataSize, err := query.GetExpectedDataBufferLength("cols")
 	require.NoError(t, err)
 	assert.Equal(t, len(buffD2), int(colsDataSize))
 
@@ -2038,9 +2646,14 @@ func TestQueryConfig(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 
+	// Create the subarray
+	subarray, err := array.NewSubarray()
+	require.NoError(t, err)
+	assert.NotNil(t, subarray)
+
 	// Read value at cell 2, 2
-	require.NoError(t, query.AddRange(0, 1, 2))
-	require.NoError(t, query.AddRange(1, 1, 2))
+	require.NoError(t, subarray.AddRange(0, MakeRange[int32](1, 2)))
+	require.NoError(t, subarray.AddRange(1, MakeRange[int32](1, 2)))
 
 	// Create configuration
 	configQuery, err := NewConfig()
@@ -2050,13 +2663,13 @@ func TestQueryConfig(t *testing.T) {
 	require.NoError(t, query.SetConfig(configQuery))
 
 	require.NoError(t, query.SetLayout(TILEDB_ROW_MAJOR))
-	_, _, err = query.SetBufferVar("a1",
-		a1OffRead, a1DataRead)
+	_, err = query.SetDataBuffer("a1", a1DataRead)
 	require.NoError(t, err)
-	assert.NotNil(t, query)
-	_, err = query.SetBuffer("rows", rows)
+	_, err = query.SetOffsetsBuffer("a1", a1OffRead)
 	require.NoError(t, err)
-	_, err = query.SetBuffer("cols", cols)
+	_, err = query.SetDataBuffer("rows", rows)
+	require.NoError(t, err)
+	_, err = query.SetDataBuffer("cols", cols)
 	require.NoError(t, err)
 
 	// Submit the query
@@ -2158,22 +2771,26 @@ func TestQueryStats(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 	require.NoError(t, query.SetLayout(TILEDB_GLOBAL_ORDER))
-	_, _, err = query.SetBufferVar("a1", a1OffWrite, a1DataWrite)
+	_, err = query.SetDataBuffer("a1", a1DataWrite)
 	require.NoError(t, err)
-	_, err = query.SetBuffer("rows", buffD1)
+	_, err = query.SetOffsetsBuffer("a1", a1OffWrite)
 	require.NoError(t, err)
-	_, err = query.SetBuffer("cols", buffD2)
+	_, err = query.SetDataBuffer("rows", buffD1)
+	require.NoError(t, err)
+	_, err = query.SetDataBuffer("cols", buffD2)
 	require.NoError(t, err)
 
 	// Check the buffer sizes
-	offsetSize, dataSize, err := query.BufferSizeVar("a1")
+	dataSize, err := query.GetExpectedDataBufferLength("a1")
+	require.NoError(t, err)
+	offsetSize, err := query.GetExpectedOffsetsBufferLength("a1")
 	require.NoError(t, err)
 	assert.Equal(t, len(a1OffWrite), int(offsetSize))
 	assert.Equal(t, len(a1DataWrite), int(dataSize))
-	rowsDataSize, err := query.BufferSize("rows")
+	rowsDataSize, err := query.GetExpectedDataBufferLength("rows")
 	require.NoError(t, err)
 	assert.Equal(t, len(buffD1), int(rowsDataSize))
-	colsDataSize, err := query.BufferSize("cols")
+	colsDataSize, err := query.GetExpectedDataBufferLength("cols")
 	require.NoError(t, err)
 	assert.Equal(t, len(buffD2), int(colsDataSize))
 
@@ -2199,15 +2816,22 @@ func TestQueryStats(t *testing.T) {
 	require.NoError(t, err)
 	assert.NotNil(t, query)
 
-	require.NoError(t, query.SetSubArray(subArray))
+	// Prepare the subarray
+	subarray, err := array.NewSubarray()
+	require.NoError(t, err)
+	assert.NotNil(t, subarray)
+	require.NoError(t, subarray.SetSubArray(subArray))
+
+	require.NoError(t, query.SetSubarray(subarray))
 	require.NoError(t, query.SetLayout(TILEDB_ROW_MAJOR))
-	offsetBufferSize, effectiveBufferSize, err := query.SetBufferVar("a1",
-		a1OffRead, a1DataRead)
+	effectiveBufferSize, err := query.SetDataBuffer("a1", a1DataRead)
+	require.NoError(t, err)
+	offsetBufferSize, err := query.SetOffsetsBuffer("a1", a1OffRead)
 	require.NoError(t, err)
 	assert.NotNil(t, query)
-	_, err = query.SetBuffer("rows", rows)
+	_, err = query.SetDataBuffer("rows", rows)
 	require.NoError(t, err)
-	_, err = query.SetBuffer("cols", cols)
+	_, err = query.SetDataBuffer("cols", cols)
 	require.NoError(t, err)
 
 	// Submit the query
@@ -2239,4 +2863,874 @@ func TestQueryStats(t *testing.T) {
 	assert.NotEmpty(t, stats, 0)
 
 	query.Free()
+}
+
+func TestSetDataBufferUnsafe(t *testing.T) {
+	// create a 1d array x[a] with a fixed length int32 attribute
+	config, err := NewConfig()
+	require.NoError(t, err)
+	tdbCtx, err := NewContext(config)
+	require.NoError(t, err)
+	dimension, err := NewDimension(tdbCtx, "x", TILEDB_INT8, []int8{1, 10}, int8(5))
+	require.NoError(t, err)
+	domain, err := NewDomain(tdbCtx)
+	require.NoError(t, err)
+	require.NoError(t, domain.AddDimensions(dimension))
+	arraySchema, err := NewArraySchema(tdbCtx, TILEDB_DENSE)
+	require.NoError(t, err)
+	attribute, err := NewAttribute(tdbCtx, "a", TILEDB_INT32)
+	require.NoError(t, err)
+	require.NoError(t, arraySchema.AddAttributes(attribute))
+	require.NoError(t, arraySchema.SetDomain(domain))
+	uri := t.TempDir()
+	array, err := NewArray(tdbCtx, uri)
+	require.NoError(t, err)
+	require.NoError(t, array.Create(arraySchema))
+	require.NoError(t, array.Close())
+
+	// open the array and write a slice
+	require.NoError(t, array.Open(TILEDB_WRITE))
+	tdbCtx, err = NewContext(config)
+	require.NoError(t, err)
+	q, err := NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err := array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](4, 7)))
+	require.NoError(t, q.SetSubarray(s))
+	dataBuffer := []int32{4, 5, 6, 7}
+	dataPtr := slicePtr(dataBuffer)
+	n, err := q.SetDataBufferUnsafe("a", dataPtr, 16)
+	require.NoError(t, err)
+	require.NotNil(t, n)
+	require.Equal(t, uint64(len(dataBuffer))*uint64(unsafe.Sizeof(dataBuffer[0])), *n)
+	require.NoError(t, q.Submit())
+	status, err := q.Status()
+	require.NoError(t, err)
+	require.Equal(t, TILEDB_COMPLETED, status)
+	require.NoError(t, array.Close())
+
+	// open the array to read the full array and verify the written cells
+	require.NoError(t, array.Open(TILEDB_READ))
+	tdbCtx, err = NewContext(config)
+	require.NoError(t, err)
+	q, err = NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err = array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](1, 10)))
+	require.NoError(t, q.SetSubarray(s))
+	dataBuffer = []int32{0, 0, 0, 0, 0, 0, 0, 0, 0, 0}
+	dataPtr = slicePtr(dataBuffer)
+	n, err = q.SetDataBufferUnsafe("a", dataPtr, 40)
+	require.NoError(t, err)
+	require.NotNil(t, n)
+	require.Equal(t, uint64(len(dataBuffer))*uint64(unsafe.Sizeof(dataBuffer[0])), *n)
+	require.NoError(t, q.Submit())
+	status, err = q.Status()
+	require.NoError(t, err)
+	require.Equal(t, TILEDB_COMPLETED, status)
+	assert.Equal(t, int32(4), dataBuffer[3])
+	assert.Equal(t, int32(5), dataBuffer[4])
+	assert.Equal(t, int32(6), dataBuffer[5])
+	assert.Equal(t, int32(7), dataBuffer[6])
+
+	// verify that GetDataBuffer works for buffers passed unsafe
+	storedBuffer, err := q.GetDataBuffer("a")
+	require.NoError(t, err)
+	storedDataBuffer, ok := storedBuffer.([]int32)
+	require.True(t, ok)
+	require.Equal(t, dataPtr, slicePtr(storedDataBuffer))
+}
+
+func TestGetDataBuffer(t *testing.T) {
+	// create a 1d array x[a] with a fixed length int32 attribute
+	config, err := NewConfig()
+	require.NoError(t, err)
+	tdbCtx, err := NewContext(config)
+	require.NoError(t, err)
+	dimension, err := NewDimension(tdbCtx, "x", TILEDB_INT8, []int8{1, 10}, int8(5))
+	require.NoError(t, err)
+	domain, err := NewDomain(tdbCtx)
+	require.NoError(t, err)
+	require.NoError(t, domain.AddDimensions(dimension))
+	arraySchema, err := NewArraySchema(tdbCtx, TILEDB_DENSE)
+	require.NoError(t, err)
+	attribute, err := NewAttribute(tdbCtx, "a", TILEDB_INT32)
+	require.NoError(t, err)
+	require.NoError(t, arraySchema.AddAttributes(attribute))
+	require.NoError(t, arraySchema.SetDomain(domain))
+	uri := t.TempDir()
+	array, err := NewArray(tdbCtx, uri)
+	require.NoError(t, err)
+	require.NoError(t, array.Create(arraySchema))
+	require.NoError(t, array.Close())
+
+	// create a write query, set the data buffer and read it back
+	require.NoError(t, array.Open(TILEDB_WRITE))
+	tdbCtx, err = NewContext(config)
+	require.NoError(t, err)
+	q, err := NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err := array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](4, 7)))
+	require.NoError(t, q.SetSubarray(s))
+	dataBuffer := []int32{4, 5, 6, 7}
+	_, err = q.SetDataBuffer("a", dataBuffer)
+	require.NoError(t, err)
+
+	storedBuffer, err := q.GetDataBuffer("a")
+	require.NoError(t, err)
+	storedDataBuffer, ok := storedBuffer.([]int32)
+	require.True(t, ok)
+	require.Equal(t, slicePtr(dataBuffer), slicePtr(storedDataBuffer))
+}
+
+func TestSetDataBuffer(t *testing.T) {
+	// create a 1d array x[a] with a fixed length int32 attribute
+	config, err := NewConfig()
+	require.NoError(t, err)
+	tdbCtx, err := NewContext(config)
+	require.NoError(t, err)
+	dimension, err := NewDimension(tdbCtx, "x", TILEDB_INT8, []int8{1, 10}, int8(5))
+	require.NoError(t, err)
+	domain, err := NewDomain(tdbCtx)
+	require.NoError(t, err)
+	require.NoError(t, domain.AddDimensions(dimension))
+	arraySchema, err := NewArraySchema(tdbCtx, TILEDB_DENSE)
+	require.NoError(t, err)
+	attribute, err := NewAttribute(tdbCtx, "a", TILEDB_INT32)
+	require.NoError(t, err)
+	require.NoError(t, arraySchema.AddAttributes(attribute))
+	require.NoError(t, arraySchema.SetDomain(domain))
+	uri := t.TempDir()
+	array, err := NewArray(tdbCtx, uri)
+	require.NoError(t, err)
+	require.NoError(t, array.Create(arraySchema))
+	require.NoError(t, array.Close())
+
+	// open the array and write a slice
+	require.NoError(t, array.Open(TILEDB_WRITE))
+	tdbCtx, err = NewContext(config)
+	require.NoError(t, err)
+	q, err := NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err := array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](4, 7)))
+	require.NoError(t, q.SetSubarray(s))
+	dataBuffer := []int32{4, 5, 6, 7}
+	np, err := q.SetDataBuffer("a", dataBuffer)
+	require.NoError(t, err)
+	require.NotNil(t, np)
+	require.Equal(t, uint64(len(dataBuffer))*uint64(unsafe.Sizeof(dataBuffer[0])), *np)
+	require.NoError(t, q.Submit())
+	status, err := q.Status()
+	require.NoError(t, err)
+	require.Equal(t, TILEDB_COMPLETED, status)
+	require.NoError(t, array.Close())
+
+	// open the array to read the full array and verify the written cells
+	require.NoError(t, array.Open(TILEDB_READ))
+	tdbCtx, err = NewContext(config)
+	require.NoError(t, err)
+	q, err = NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err = array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](1, 10)))
+	require.NoError(t, q.SetSubarray(s))
+	dataBuffer = []int32{0, 0, 0, 0, 0, 0, 0, 0, 0, 0}
+	np, err = q.SetDataBuffer("a", dataBuffer)
+	require.NoError(t, err)
+	require.NotNil(t, np)
+	require.Equal(t, uint64(len(dataBuffer))*uint64(unsafe.Sizeof(dataBuffer[0])), *np)
+	require.NoError(t, q.Submit())
+	status, err = q.Status()
+	require.NoError(t, err)
+	require.Equal(t, TILEDB_COMPLETED, status)
+	require.NoError(t, array.Close())
+	assert.Equal(t, int32(4), dataBuffer[3])
+	assert.Equal(t, int32(5), dataBuffer[4])
+	assert.Equal(t, int32(6), dataBuffer[5])
+	assert.Equal(t, int32(7), dataBuffer[6])
+}
+
+func TestGetExpectedDataBufferLength(t *testing.T) {
+	// create a 1d array x[a] with a fixed length int32 attribute
+	config, err := NewConfig()
+	require.NoError(t, err)
+	tdbCtx, err := NewContext(config)
+	require.NoError(t, err)
+	dimension, err := NewDimension(tdbCtx, "x", TILEDB_INT8, []int8{1, 10}, int8(5))
+	require.NoError(t, err)
+	domain, err := NewDomain(tdbCtx)
+	require.NoError(t, err)
+	require.NoError(t, domain.AddDimensions(dimension))
+	arraySchema, err := NewArraySchema(tdbCtx, TILEDB_DENSE)
+	require.NoError(t, err)
+	attribute, err := NewAttribute(tdbCtx, "a", TILEDB_INT32)
+	require.NoError(t, err)
+	require.NoError(t, arraySchema.AddAttributes(attribute))
+	require.NoError(t, arraySchema.SetDomain(domain))
+	uri := t.TempDir()
+	array, err := NewArray(tdbCtx, uri)
+	require.NoError(t, err)
+	require.NoError(t, array.Create(arraySchema))
+
+	require.NoError(t, array.Open(TILEDB_READ))
+	q, err := NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err := array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](4, 7)))
+	require.NoError(t, q.SetSubarray(s))
+	dataBuffer := []int32{0, 0, 0, 0}
+	_, err = q.SetDataBuffer("a", dataBuffer)
+	require.NoError(t, err)
+
+	t.Run("ProperQuery", func(t *testing.T) {
+		storedBuffer, err := q.GetDataBuffer("a")
+		require.NoError(t, err)
+		require.NotNil(t, storedBuffer)
+		siz, err := q.GetExpectedDataBufferLength("a")
+		require.NoError(t, err)
+		require.Equal(t, uint64(4), siz)
+	})
+
+	t.Run("SerializedClientSideQuery", func(t *testing.T) {
+		bf, err := SerializeQuery(q, TILEDB_CAPNP, true)
+		require.NoError(t, err)
+		require.NotNil(t, bf)
+		buf, err := bf.Flatten()
+		require.NoError(t, err)
+
+		dq, err := NewQuery(tdbCtx, array)
+		require.NoError(t, err)
+		_, err = dq.SetDataBuffer("a", []int32{0, 0, 0, 0})
+		require.NoError(t, err)
+		err = DeserializeQuery(dq, buf, TILEDB_CAPNP, true)
+		require.NoError(t, err)
+
+		storedBuffer, err := dq.GetDataBuffer("a")
+		require.NoError(t, err)
+		require.NotNil(t, storedBuffer)
+		siz, err := dq.GetExpectedDataBufferLength("a")
+		require.NoError(t, err)
+		require.Equal(t, uint64(4), siz)
+	})
+
+	t.Run("SerializedServerSideQuery", func(t *testing.T) {
+		bf, err := SerializeQuery(q, TILEDB_CAPNP, false)
+		require.NoError(t, err)
+		require.NotNil(t, bf)
+		buf, err := bf.Flatten()
+		require.NoError(t, err)
+
+		dq, err := NewQuery(tdbCtx, array)
+		require.NoError(t, err)
+		err = DeserializeQuery(dq, buf, TILEDB_CAPNP, false)
+		require.NoError(t, err)
+
+		storedBuffer, err := dq.GetDataBuffer("a")
+		require.NoError(t, err)
+		require.Nil(t, storedBuffer)
+		siz, err := dq.GetExpectedDataBufferLength("a")
+		require.NoError(t, err)
+		require.Equal(t, uint64(4), siz)
+	})
+}
+
+func TestSetValidityBufferUnsafe(t *testing.T) {
+	// create a 1d array x[a] with a fixed length nullable int32 attribute
+	config, err := NewConfig()
+	require.NoError(t, err)
+	tdbCtx, err := NewContext(config)
+	require.NoError(t, err)
+	dimension, err := NewDimension(tdbCtx, "x", TILEDB_INT8, []int8{1, 10}, int8(5))
+	require.NoError(t, err)
+	domain, err := NewDomain(tdbCtx)
+	require.NoError(t, err)
+	require.NoError(t, domain.AddDimensions(dimension))
+	arraySchema, err := NewArraySchema(tdbCtx, TILEDB_DENSE)
+	require.NoError(t, err)
+	attribute, err := NewAttribute(tdbCtx, "a", TILEDB_INT32)
+	require.NoError(t, err)
+	require.NoError(t, attribute.SetNullable(true))
+	require.NoError(t, arraySchema.AddAttributes(attribute))
+	require.NoError(t, arraySchema.SetDomain(domain))
+	uri := t.TempDir()
+	array, err := NewArray(tdbCtx, uri)
+	require.NoError(t, err)
+	require.NoError(t, array.Create(arraySchema))
+	require.NoError(t, array.Close())
+
+	// open the array and write a slice
+	require.NoError(t, array.Open(TILEDB_WRITE))
+	tdbCtx, err = NewContext(config)
+	require.NoError(t, err)
+	q, err := NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err := array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](4, 7)))
+	require.NoError(t, q.SetSubarray(s))
+	dataBuffer := []int32{4, 5, 6, 7}
+	_, err = q.SetDataBuffer("a", dataBuffer)
+	require.NoError(t, err)
+	validityBuffer := []uint8{1, 1, 0, 1}
+	validityPtr := slicePtr(validityBuffer)
+	n, err := q.SetValidityBufferUnsafe("a", validityPtr, 4)
+	require.NoError(t, err)
+	require.NotNil(t, n)
+	require.Equal(t, uint64(len(validityBuffer))*uint64(unsafe.Sizeof(validityBuffer[0])), *n)
+	require.NoError(t, q.Submit())
+	status, err := q.Status()
+	require.NoError(t, err)
+	require.Equal(t, TILEDB_COMPLETED, status)
+	require.NoError(t, array.Close())
+
+	// open the array to read the full array and verify the written cells
+	require.NoError(t, array.Open(TILEDB_READ))
+	tdbCtx, err = NewContext(config)
+	require.NoError(t, err)
+	q, err = NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err = array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](1, 10)))
+	require.NoError(t, q.SetSubarray(s))
+	dataBuffer = []int32{0, 0, 0, 0, 0, 0, 0, 0, 0, 0}
+	_, err = q.SetDataBuffer("a", dataBuffer)
+	require.NoError(t, err)
+	validityBuffer = []uint8{0, 0, 0, 0, 0, 0, 0, 0, 0, 0}
+	validityPtr = slicePtr(validityBuffer)
+	n, err = q.SetValidityBufferUnsafe("a", validityPtr, 10)
+	require.NoError(t, err)
+	require.NotNil(t, n)
+	require.Equal(t, uint64(len(validityBuffer))*uint64(unsafe.Sizeof(validityBuffer[0])), *n)
+	require.NoError(t, q.Submit())
+	status, err = q.Status()
+	require.NoError(t, err)
+	require.Equal(t, TILEDB_COMPLETED, status)
+	assert.Equal(t, int32(4), dataBuffer[3])
+	assert.Equal(t, int32(5), dataBuffer[4])
+	assert.Equal(t, int32(6), dataBuffer[5])
+	assert.Equal(t, int32(7), dataBuffer[6])
+
+	// verify that GetDataBuffer works for buffers passed unsafe
+	storedValidityBuffer, err := q.GetValidityBuffer("a")
+	require.NoError(t, err)
+	require.Equal(t, validityPtr, slicePtr(storedValidityBuffer))
+}
+
+func TestGetValidityBuffer(t *testing.T) {
+	// create a 1d array x[a] with a fixed length nullable int32 attribute
+	config, err := NewConfig()
+	require.NoError(t, err)
+	tdbCtx, err := NewContext(config)
+	require.NoError(t, err)
+	dimension, err := NewDimension(tdbCtx, "x", TILEDB_INT8, []int8{1, 10}, int8(5))
+	require.NoError(t, err)
+	domain, err := NewDomain(tdbCtx)
+	require.NoError(t, err)
+	require.NoError(t, domain.AddDimensions(dimension))
+	arraySchema, err := NewArraySchema(tdbCtx, TILEDB_DENSE)
+	require.NoError(t, err)
+	attribute, err := NewAttribute(tdbCtx, "a", TILEDB_INT32)
+	require.NoError(t, err)
+	require.NoError(t, attribute.SetNullable(true))
+	require.NoError(t, arraySchema.AddAttributes(attribute))
+	require.NoError(t, arraySchema.SetDomain(domain))
+	uri := t.TempDir()
+	array, err := NewArray(tdbCtx, uri)
+	require.NoError(t, err)
+	require.NoError(t, array.Create(arraySchema))
+	require.NoError(t, array.Close())
+
+	// create a write query, set the validity buffer and read it back
+	require.NoError(t, array.Open(TILEDB_WRITE))
+	tdbCtx, err = NewContext(config)
+	require.NoError(t, err)
+	q, err := NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err := array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](4, 7)))
+	require.NoError(t, q.SetSubarray(s))
+	validityBuffer := []uint8{1, 1, 0, 1}
+	_, err = q.SetValidityBuffer("a", validityBuffer)
+	require.NoError(t, err)
+
+	storedValidityBuffer, err := q.GetValidityBuffer("a")
+	require.NoError(t, err)
+	require.Equal(t, slicePtr(validityBuffer), slicePtr(storedValidityBuffer))
+}
+
+func TestSetValidityBuffer(t *testing.T) {
+	// create a 1d array x[a] with a fixed length nullable int32 attribute
+	config, err := NewConfig()
+	require.NoError(t, err)
+	tdbCtx, err := NewContext(config)
+	require.NoError(t, err)
+	dimension, err := NewDimension(tdbCtx, "x", TILEDB_INT8, []int8{1, 10}, int8(5))
+	require.NoError(t, err)
+	domain, err := NewDomain(tdbCtx)
+	require.NoError(t, err)
+	require.NoError(t, domain.AddDimensions(dimension))
+	arraySchema, err := NewArraySchema(tdbCtx, TILEDB_DENSE)
+	require.NoError(t, err)
+	attribute, err := NewAttribute(tdbCtx, "a", TILEDB_INT32)
+	require.NoError(t, err)
+	require.NoError(t, attribute.SetNullable(true))
+	require.NoError(t, arraySchema.AddAttributes(attribute))
+	require.NoError(t, arraySchema.SetDomain(domain))
+	uri := t.TempDir()
+	array, err := NewArray(tdbCtx, uri)
+	require.NoError(t, err)
+	require.NoError(t, array.Create(arraySchema))
+	require.NoError(t, array.Close())
+
+	// open the array and write a slice
+	require.NoError(t, array.Open(TILEDB_WRITE))
+	tdbCtx, err = NewContext(config)
+	require.NoError(t, err)
+	q, err := NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err := array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](4, 7)))
+	require.NoError(t, q.SetSubarray(s))
+	dataBuffer := []int32{4, 5, 6, 7}
+	np, err := q.SetDataBuffer("a", dataBuffer)
+	require.NoError(t, err)
+	require.NotNil(t, np)
+	require.Equal(t, uint64(len(dataBuffer))*uint64(unsafe.Sizeof(dataBuffer[0])), *np)
+	validityBuffer := []uint8{1, 1, 0, 1}
+	vnp, err := q.SetValidityBuffer("a", validityBuffer)
+	require.NoError(t, err)
+	require.NotNil(t, np)
+	require.Equal(t, uint64(len(validityBuffer))*uint64(unsafe.Sizeof(validityBuffer[0])), *vnp)
+	require.NoError(t, q.Submit())
+	status, err := q.Status()
+	require.NoError(t, err)
+	require.Equal(t, TILEDB_COMPLETED, status)
+	require.NoError(t, array.Close())
+
+	// open the array to read the full array and verify the written cells
+	require.NoError(t, array.Open(TILEDB_READ))
+	tdbCtx, err = NewContext(config)
+	require.NoError(t, err)
+	q, err = NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err = array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](1, 10)))
+	require.NoError(t, q.SetSubarray(s))
+	dataBuffer = []int32{0, 0, 0, 0, 0, 0, 0, 0, 0, 0}
+	np, err = q.SetDataBuffer("a", dataBuffer)
+	require.NoError(t, err)
+	require.NotNil(t, np)
+	require.Equal(t, uint64(len(dataBuffer))*uint64(unsafe.Sizeof(dataBuffer[0])), *np)
+	validityBuffer = []uint8{2, 2, 2, 2, 2, 2, 2, 2, 2, 2}
+	vnp, err = q.SetValidityBuffer("a", validityBuffer)
+	require.NoError(t, err)
+	require.NotNil(t, np)
+	require.Equal(t, uint64(len(validityBuffer))*uint64(unsafe.Sizeof(validityBuffer[0])), *vnp)
+	require.NoError(t, q.Submit())
+	status, err = q.Status()
+	require.NoError(t, err)
+	require.Equal(t, TILEDB_COMPLETED, status)
+	require.NoError(t, array.Close())
+	assert.Equal(t, []uint8{0, 0, 0, 1, 1, 0, 1, 0, 0, 0}, validityBuffer)
+}
+
+func TestGetExpectedValidityBufferLength(t *testing.T) {
+	// create a 1d array x[a] with a fixed length nullable int32 attribute
+	config, err := NewConfig()
+	require.NoError(t, err)
+	tdbCtx, err := NewContext(config)
+	require.NoError(t, err)
+	dimension, err := NewDimension(tdbCtx, "x", TILEDB_INT8, []int8{1, 10}, int8(5))
+	require.NoError(t, err)
+	domain, err := NewDomain(tdbCtx)
+	require.NoError(t, err)
+	require.NoError(t, domain.AddDimensions(dimension))
+	arraySchema, err := NewArraySchema(tdbCtx, TILEDB_DENSE)
+	require.NoError(t, err)
+	attribute, err := NewAttribute(tdbCtx, "a", TILEDB_INT32)
+	require.NoError(t, err)
+	require.NoError(t, attribute.SetNullable(true))
+	require.NoError(t, arraySchema.AddAttributes(attribute))
+	require.NoError(t, arraySchema.SetDomain(domain))
+	uri := t.TempDir()
+	array, err := NewArray(tdbCtx, uri)
+	require.NoError(t, err)
+	require.NoError(t, array.Create(arraySchema))
+	require.NoError(t, array.Close())
+
+	require.NoError(t, array.Open(TILEDB_READ))
+	q, err := NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err := array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](4, 7)))
+	require.NoError(t, q.SetSubarray(s))
+	dataBuffer := []int32{4, 5, 6, 7}
+	_, err = q.SetDataBuffer("a", dataBuffer)
+	require.NoError(t, err)
+	validityBuffer := []uint8{1, 1, 0, 1}
+	_, err = q.SetValidityBuffer("a", validityBuffer)
+	require.NoError(t, err)
+
+	t.Run("ProperQuery", func(t *testing.T) {
+		storedBuffer, err := q.GetValidityBuffer("a")
+		require.NoError(t, err)
+		require.NotNil(t, storedBuffer)
+		siz, err := q.GetExpectedValidityBufferLength("a")
+		require.NoError(t, err)
+		require.Equal(t, uint64(4), siz)
+	})
+
+	t.Run("SerializedClientSideQuery", func(t *testing.T) {
+		bf, err := SerializeQuery(q, TILEDB_CAPNP, true)
+		require.NoError(t, err)
+		require.NotNil(t, bf)
+		buf, err := bf.Flatten()
+		require.NoError(t, err)
+
+		dq, err := NewQuery(tdbCtx, array)
+		require.NoError(t, err)
+		_, err = dq.SetDataBuffer("a", []int32{0, 0, 0, 0})
+		require.NoError(t, err)
+		_, err = dq.SetValidityBuffer("a", []uint8{0, 0, 0, 0})
+		require.NoError(t, err)
+		err = DeserializeQuery(dq, buf, TILEDB_CAPNP, true)
+		require.NoError(t, err)
+
+		storedBuffer, err := dq.GetValidityBuffer("a")
+		require.NoError(t, err)
+		require.NotNil(t, storedBuffer)
+		siz, err := dq.GetExpectedValidityBufferLength("a")
+		require.NoError(t, err)
+		require.Equal(t, uint64(4), siz)
+	})
+
+	t.Run("SerializedServerSideQuery", func(t *testing.T) {
+		bf, err := SerializeQuery(q, TILEDB_CAPNP, false)
+		require.NoError(t, err)
+		require.NotNil(t, bf)
+		buf, err := bf.Flatten()
+		require.NoError(t, err)
+
+		dq, err := NewQuery(tdbCtx, array)
+		require.NoError(t, err)
+		err = DeserializeQuery(dq, buf, TILEDB_CAPNP, false)
+		require.NoError(t, err)
+
+		storedBuffer, err := dq.GetValidityBuffer("a")
+		require.NoError(t, err)
+		require.Nil(t, storedBuffer)
+		siz, err := dq.GetExpectedValidityBufferLength("a")
+		require.NoError(t, err)
+		require.Equal(t, uint64(4), siz)
+	})
+}
+
+func TestSetOffsetsBufferUnsafe(t *testing.T) {
+	// create a 1d array x[a] with a var length attribute
+	config, err := NewConfig()
+	require.NoError(t, err)
+	tdbCtx, err := NewContext(config)
+	require.NoError(t, err)
+	dimension, err := NewDimension(tdbCtx, "x", TILEDB_INT8, []int8{1, 10}, int8(5))
+	require.NoError(t, err)
+	domain, err := NewDomain(tdbCtx)
+	require.NoError(t, err)
+	require.NoError(t, domain.AddDimensions(dimension))
+	arraySchema, err := NewArraySchema(tdbCtx, TILEDB_DENSE)
+	require.NoError(t, err)
+	attribute, err := NewAttribute(tdbCtx, "a", TILEDB_STRING_ASCII)
+	require.NoError(t, err)
+	require.NoError(t, attribute.SetCellValNum(TILEDB_VAR_NUM))
+	require.NoError(t, arraySchema.AddAttributes(attribute))
+	require.NoError(t, arraySchema.SetDomain(domain))
+	uri := t.TempDir()
+	array, err := NewArray(tdbCtx, uri)
+	require.NoError(t, err)
+	require.NoError(t, array.Create(arraySchema))
+	require.NoError(t, array.Close())
+
+	// open the array and write a slice
+	require.NoError(t, array.Open(TILEDB_WRITE))
+	tdbCtx, err = NewContext(config)
+	require.NoError(t, err)
+	q, err := NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err := array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](4, 7)))
+	require.NoError(t, q.SetSubarray(s))
+	dataBuffer := []byte("HelloWorldFromTiledb")
+	np, err := q.SetDataBuffer("a", dataBuffer)
+	require.NoError(t, err)
+	require.NotNil(t, np)
+	require.Equal(t, uint64(len(dataBuffer))*uint64(unsafe.Sizeof(dataBuffer[0])), *np)
+	offsetsBuffer := []uint64{0, 5, 10, 14}
+	offsetsPtr := slicePtr(offsetsBuffer)
+	vnp, err := q.SetOffsetsBufferUnsafe("a", offsetsPtr, 32)
+	require.NoError(t, err)
+	require.NotNil(t, np)
+	require.Equal(t, uint64(len(offsetsBuffer))*uint64(unsafe.Sizeof(offsetsBuffer[0])), *vnp)
+	require.NoError(t, q.Submit())
+	status, err := q.Status()
+	require.NoError(t, err)
+	require.Equal(t, TILEDB_COMPLETED, status)
+	require.NoError(t, array.Close())
+
+	// open the array to read the full array and verify the written cells
+	require.NoError(t, array.Open(TILEDB_READ))
+	tdbCtx, err = NewContext(config)
+	require.NoError(t, err)
+	q, err = NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err = array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](1, 10)))
+	require.NoError(t, q.SetSubarray(s))
+	dataBuffer = make([]byte, 40)
+	np, err = q.SetDataBuffer("a", dataBuffer)
+	require.NoError(t, err)
+	require.NotNil(t, np)
+	require.Equal(t, uint64(len(dataBuffer))*uint64(unsafe.Sizeof(dataBuffer[0])), *np)
+	offsetsBuffer = make([]uint64, 10)
+	offsetsPtr = slicePtr(offsetsBuffer)
+	vnp, err = q.SetOffsetsBufferUnsafe("a", offsetsPtr, 80)
+	require.NoError(t, err)
+	require.NotNil(t, vnp)
+	require.Equal(t, uint64(len(offsetsBuffer))*uint64(unsafe.Sizeof(offsetsBuffer[0])), *vnp)
+	require.NoError(t, q.Submit())
+	status, err = q.Status()
+	require.NoError(t, err)
+	require.Equal(t, TILEDB_COMPLETED, status)
+	assert.Equal(t, []uint64{0, 1, 2, 3, 8, 13, 17, 23, 24, 25}, offsetsBuffer)
+
+	// verify that GetOffsetsBuffer works for buffers passed unsafe
+	storedOffsetsBuffer, err := q.GetOffsetsBuffer("a")
+	require.NoError(t, err)
+	require.Equal(t, offsetsPtr, slicePtr(storedOffsetsBuffer))
+}
+
+func TestGetOffsetsBuffer(t *testing.T) {
+	// create a 1d array x[a] with a var length attribute
+	config, err := NewConfig()
+	require.NoError(t, err)
+	tdbCtx, err := NewContext(config)
+	require.NoError(t, err)
+	dimension, err := NewDimension(tdbCtx, "x", TILEDB_INT8, []int8{1, 10}, int8(5))
+	require.NoError(t, err)
+	domain, err := NewDomain(tdbCtx)
+	require.NoError(t, err)
+	require.NoError(t, domain.AddDimensions(dimension))
+	arraySchema, err := NewArraySchema(tdbCtx, TILEDB_DENSE)
+	require.NoError(t, err)
+	attribute, err := NewAttribute(tdbCtx, "a", TILEDB_STRING_ASCII)
+	require.NoError(t, err)
+	require.NoError(t, attribute.SetCellValNum(TILEDB_VAR_NUM))
+	require.NoError(t, arraySchema.AddAttributes(attribute))
+	require.NoError(t, arraySchema.SetDomain(domain))
+	uri := t.TempDir()
+	array, err := NewArray(tdbCtx, uri)
+	require.NoError(t, err)
+	require.NoError(t, array.Create(arraySchema))
+	require.NoError(t, array.Close())
+
+	// create a write query, set the validity buffer and read it back
+	require.NoError(t, array.Open(TILEDB_WRITE))
+	tdbCtx, err = NewContext(config)
+	require.NoError(t, err)
+	q, err := NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err := array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](4, 7)))
+	require.NoError(t, q.SetSubarray(s))
+	offsetsBuffer := []uint64{0, 2, 4, 8}
+	_, err = q.SetOffsetsBuffer("a", offsetsBuffer)
+	require.NoError(t, err)
+
+	storedOffsetsBuffer, err := q.GetOffsetsBuffer("a")
+	require.NoError(t, err)
+	require.Equal(t, slicePtr(offsetsBuffer), slicePtr(storedOffsetsBuffer))
+}
+
+func TestSetOffsetsBuffer(t *testing.T) {
+	// create a 1d array x[a] with a var length attribute
+	config, err := NewConfig()
+	require.NoError(t, err)
+	tdbCtx, err := NewContext(config)
+	require.NoError(t, err)
+	dimension, err := NewDimension(tdbCtx, "x", TILEDB_INT8, []int8{1, 10}, int8(5))
+	require.NoError(t, err)
+	domain, err := NewDomain(tdbCtx)
+	require.NoError(t, err)
+	require.NoError(t, domain.AddDimensions(dimension))
+	arraySchema, err := NewArraySchema(tdbCtx, TILEDB_DENSE)
+	require.NoError(t, err)
+	attribute, err := NewAttribute(tdbCtx, "a", TILEDB_STRING_ASCII)
+	require.NoError(t, err)
+	require.NoError(t, attribute.SetCellValNum(TILEDB_VAR_NUM))
+	require.NoError(t, arraySchema.AddAttributes(attribute))
+	require.NoError(t, arraySchema.SetDomain(domain))
+	uri := t.TempDir()
+	array, err := NewArray(tdbCtx, uri)
+	require.NoError(t, err)
+	require.NoError(t, array.Create(arraySchema))
+	require.NoError(t, array.Close())
+
+	// open the array and write a slice
+	require.NoError(t, array.Open(TILEDB_WRITE))
+	tdbCtx, err = NewContext(config)
+	require.NoError(t, err)
+	q, err := NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err := array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](4, 7)))
+	require.NoError(t, q.SetSubarray(s))
+	dataBuffer := []byte("HelloWorldFromTiledb")
+	np, err := q.SetDataBuffer("a", dataBuffer)
+	require.NoError(t, err)
+	require.NotNil(t, np)
+	require.Equal(t, uint64(len(dataBuffer))*uint64(unsafe.Sizeof(dataBuffer[0])), *np)
+	offsetsBuffer := []uint64{0, 5, 10, 14}
+	vnp, err := q.SetOffsetsBuffer("a", offsetsBuffer)
+	require.NoError(t, err)
+	require.NotNil(t, np)
+	require.Equal(t, uint64(len(offsetsBuffer))*uint64(unsafe.Sizeof(offsetsBuffer[0])), *vnp)
+	require.NoError(t, q.Submit())
+	status, err := q.Status()
+	require.NoError(t, err)
+	require.Equal(t, TILEDB_COMPLETED, status)
+	require.NoError(t, array.Close())
+
+	// open the array to read the full array and verify the written cells
+	require.NoError(t, array.Open(TILEDB_READ))
+	tdbCtx, err = NewContext(config)
+	require.NoError(t, err)
+	q, err = NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err = array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](1, 10)))
+	require.NoError(t, q.SetSubarray(s))
+	dataBuffer = make([]byte, 40)
+	np, err = q.SetDataBuffer("a", dataBuffer)
+	require.NoError(t, err)
+	require.NotNil(t, np)
+	require.Equal(t, uint64(len(dataBuffer))*uint64(unsafe.Sizeof(dataBuffer[0])), *np)
+	offsetsBuffer = make([]uint64, 10)
+	vnp, err = q.SetOffsetsBuffer("a", offsetsBuffer)
+	require.NoError(t, err)
+	require.NotNil(t, np)
+	require.Equal(t, uint64(len(offsetsBuffer))*uint64(unsafe.Sizeof(offsetsBuffer[0])), *vnp)
+	require.NoError(t, q.Submit())
+	status, err = q.Status()
+	require.NoError(t, err)
+	require.Equal(t, TILEDB_COMPLETED, status)
+	require.NoError(t, array.Close())
+	assert.Equal(t, []uint64{0, 1, 2, 3, 8, 13, 17, 23, 24, 25}, offsetsBuffer)
+}
+
+func TestGetExpectedOffsetsBufferLength(t *testing.T) {
+	// create a 1d array x[a] with a var length attribute
+	config, err := NewConfig()
+	require.NoError(t, err)
+	tdbCtx, err := NewContext(config)
+	require.NoError(t, err)
+	dimension, err := NewDimension(tdbCtx, "x", TILEDB_INT8, []int8{1, 10}, int8(5))
+	require.NoError(t, err)
+	domain, err := NewDomain(tdbCtx)
+	require.NoError(t, err)
+	require.NoError(t, domain.AddDimensions(dimension))
+	arraySchema, err := NewArraySchema(tdbCtx, TILEDB_DENSE)
+	require.NoError(t, err)
+	attribute, err := NewAttribute(tdbCtx, "a", TILEDB_STRING_ASCII)
+	require.NoError(t, err)
+	require.NoError(t, attribute.SetCellValNum(TILEDB_VAR_NUM))
+	require.NoError(t, arraySchema.AddAttributes(attribute))
+	require.NoError(t, arraySchema.SetDomain(domain))
+	uri := t.TempDir()
+	array, err := NewArray(tdbCtx, uri)
+	require.NoError(t, err)
+	require.NoError(t, array.Create(arraySchema))
+	require.NoError(t, array.Close())
+
+	require.NoError(t, array.Open(TILEDB_READ))
+	q, err := NewQuery(tdbCtx, array)
+	require.NoError(t, err)
+	s, err := array.NewSubarray()
+	require.NoError(t, err)
+	require.NoError(t, s.AddRangeByName("x", MakeRange[int8](4, 7)))
+	require.NoError(t, q.SetSubarray(s))
+	dataBuffer := make([]byte, 40)
+	_, err = q.SetDataBuffer("a", dataBuffer)
+	require.NoError(t, err)
+	offsetsBuffer := []uint64{0, 0, 0, 0}
+	_, err = q.SetOffsetsBuffer("a", offsetsBuffer)
+	require.NoError(t, err)
+
+	t.Run("ProperQuery", func(t *testing.T) {
+		storedBuffer, err := q.GetOffsetsBuffer("a")
+		require.NoError(t, err)
+		require.NotNil(t, storedBuffer)
+		siz, err := q.GetExpectedOffsetsBufferLength("a")
+		require.NoError(t, err)
+		require.Equal(t, uint64(4), siz)
+	})
+
+	t.Run("SerializedClientSideQuery", func(t *testing.T) {
+		bf, err := SerializeQuery(q, TILEDB_CAPNP, true)
+		require.NoError(t, err)
+		require.NotNil(t, bf)
+		buf, err := bf.Flatten()
+		require.NoError(t, err)
+
+		dq, err := NewQuery(tdbCtx, array)
+		require.NoError(t, err)
+		_, err = dq.SetDataBuffer("a", make([]byte, 40))
+		require.NoError(t, err)
+		_, err = dq.SetOffsetsBuffer("a", []uint64{0, 0, 0, 0})
+		require.NoError(t, err)
+		err = DeserializeQuery(dq, buf, TILEDB_CAPNP, true)
+		require.NoError(t, err)
+
+		storedBuffer, err := dq.GetOffsetsBuffer("a")
+		require.NoError(t, err)
+		require.NotNil(t, storedBuffer)
+		siz, err := dq.GetExpectedOffsetsBufferLength("a")
+		require.NoError(t, err)
+		require.Equal(t, uint64(4), siz)
+	})
+
+	t.Run("SerializedServerSideQuery", func(t *testing.T) {
+		bf, err := SerializeQuery(q, TILEDB_CAPNP, false)
+		require.NoError(t, err)
+		require.NotNil(t, bf)
+		buf, err := bf.Flatten()
+		require.NoError(t, err)
+
+		dq, err := NewQuery(tdbCtx, array)
+		require.NoError(t, err)
+		err = DeserializeQuery(dq, buf, TILEDB_CAPNP, false)
+		require.NoError(t, err)
+
+		storedBuffer, err := dq.GetOffsetsBuffer("a")
+		require.NoError(t, err)
+		require.Nil(t, storedBuffer)
+		siz, err := dq.GetExpectedOffsetsBufferLength("a")
+		require.NoError(t, err)
+		require.Equal(t, uint64(4), siz)
+	})
 }
