@@ -1,8 +1,6 @@
 package tiledb
 
 /*
-#cgo LDFLAGS: -ltiledb
-#cgo linux LDFLAGS: -ldl
 #include <tiledb/tiledb.h>
 #include <tiledb/tiledb_serialization.h>
 #include <stdio.h>
@@ -12,12 +10,11 @@ import "C"
 
 import (
 	"fmt"
-	"reflect"
 	"runtime"
 	"unsafe"
 )
 
-// SerializeArraySchema serializes an array schema
+// SerializeArraySchema serializes an array schema.
 func SerializeArraySchema(schema *ArraySchema, serializationType SerializationType, clientSide bool) ([]byte, error) {
 	var cClientSide C.int32_t
 	if clientSide {
@@ -27,10 +24,7 @@ func SerializeArraySchema(schema *ArraySchema, serializationType SerializationTy
 	}
 
 	buffer := Buffer{context: schema.context}
-	// Set finalizer for free C pointer on gc
-	runtime.SetFinalizer(&buffer, func(buffer *Buffer) {
-		buffer.Free()
-	})
+	freeOnGC(&buffer)
 
 	ret := C.tiledb_serialize_array_schema(schema.context.tiledbContext, schema.tiledbArraySchema, C.tiledb_serialization_type_t(serializationType), cClientSide, &buffer.tiledbBuffer)
 	if ret != C.TILEDB_OK {
@@ -40,7 +34,7 @@ func SerializeArraySchema(schema *ArraySchema, serializationType SerializationTy
 	return buffer.Serialize(serializationType)
 }
 
-// DeserializeArraySchema deserializes a new array schema from the given buffer
+// DeserializeArraySchema deserializes a new array schema from the given buffer.
 func DeserializeArraySchema(buffer *Buffer, serializationType SerializationType, clientSide bool) (*ArraySchema, error) {
 	schema := ArraySchema{context: buffer.context}
 
@@ -56,18 +50,15 @@ func DeserializeArraySchema(buffer *Buffer, serializationType SerializationType,
 		return nil, fmt.Errorf("Error deserializing array schema: %s", schema.context.LastError())
 	}
 
-	// Set finalizer for free C pointer on gc
 	// This needs to happen *after* the tiledb_deserialize_array_schema call
 	// because that may leave the arraySchema with a non-nil pointer
 	// to already-freed memory.
-	runtime.SetFinalizer(&schema, func(arraySchema *ArraySchema) {
-		arraySchema.Free()
-	})
+	freeOnGC(&schema)
 
 	return &schema, nil
 }
 
-// SerializeArrayNonEmptyDomain gets and serializes the array nonempty domain
+// SerializeArrayNonEmptyDomain gets and serializes the array nonempty domain.
 func SerializeArrayNonEmptyDomain(a *Array, serializationType SerializationType) ([]byte, error) {
 	schema, err := a.Schema()
 	if err != nil {
@@ -89,19 +80,16 @@ func SerializeArrayNonEmptyDomain(a *Array, serializationType SerializationType)
 
 	var isEmpty C.int32_t
 	tmpDomain := make([]uint8, subarraySize)
-	ret := C.tiledb_array_get_non_empty_domain(a.context.tiledbContext, a.tiledbArray, unsafe.Pointer(&tmpDomain[0]), &isEmpty)
+	ret := C.tiledb_array_get_non_empty_domain(a.context.tiledbContext, a.tiledbArray, slicePtr(tmpDomain), &isEmpty)
 	if ret != C.TILEDB_OK {
 		return nil, fmt.Errorf("Error serializing array nonempty domain: %s", a.context.LastError())
 	}
 
 	buffer := Buffer{context: schema.context}
-	// Set finalizer for free C pointer on gc
-	runtime.SetFinalizer(&buffer, func(buffer *Buffer) {
-		buffer.Free()
-	})
+	freeOnGC(&buffer)
 
 	var cClientSide = C.int32_t(0) // Currently this parameter is unused in libtiledb
-	ret = C.tiledb_serialize_array_nonempty_domain(a.context.tiledbContext, a.tiledbArray, unsafe.Pointer(&tmpDomain[0]), isEmpty, C.tiledb_serialization_type_t(serializationType), cClientSide, &buffer.tiledbBuffer)
+	ret = C.tiledb_serialize_array_nonempty_domain(a.context.tiledbContext, a.tiledbArray, slicePtr(tmpDomain), isEmpty, C.tiledb_serialization_type_t(serializationType), cClientSide, &buffer.tiledbBuffer)
 	if ret != C.TILEDB_OK {
 		return nil, fmt.Errorf("Error serializing array nonempty domain: %s", a.context.LastError())
 	}
@@ -109,7 +97,7 @@ func SerializeArrayNonEmptyDomain(a *Array, serializationType SerializationType)
 	return buffer.Serialize(serializationType)
 }
 
-// DeserializeArrayNonEmptyDomain deserializes an array nonempty domain
+// DeserializeArrayNonEmptyDomain deserializes an array nonempty domain.
 func DeserializeArrayNonEmptyDomain(a *Array, buffer *Buffer, serializationType SerializationType) ([]NonEmptyDomain, bool, error) {
 	schema, err := a.Schema()
 	if err != nil {
@@ -144,86 +132,58 @@ func DeserializeArrayNonEmptyDomain(a *Array, buffer *Buffer, serializationType 
 		return nil, true, nil
 	}
 
-	tmpDomainType := reflect.TypeOf(tmpDomain).Elem().Kind()
-	nonEmptyDomains := make([]NonEmptyDomain, 0)
-	for i := uint(0); i < ndims; i++ {
-		dimension, err := domain.DimensionFromIndex(i)
+	nonEmptyDomains := make([]NonEmptyDomain, ndims)
+	for i := range nonEmptyDomains {
+		dimension, err := domain.DimensionFromIndex(uint(i))
 		if err != nil {
 			return nil, false, err
 		}
 
 		var nonEmptyDomain *NonEmptyDomain
 
-		switch tmpDomainType {
-		case reflect.Int:
-			tmpSubArray := tmpDomain.([]int)
-			tmpDimension := []int{tmpSubArray[(2 * i)], tmpSubArray[(2*i)+1]}
-			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDimension)
-		case reflect.Int8:
-			tmpSubArray := tmpDomain.([]int8)
-			tmpDimension := []int8{tmpSubArray[(2 * i)], tmpSubArray[(2*i)+1]}
-			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDimension)
-		case reflect.Int16:
-			tmpSubArray := tmpDomain.([]int16)
-			tmpDimension := []int16{tmpSubArray[(2 * i)], tmpSubArray[(2*i)+1]}
-			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDimension)
-		case reflect.Int32:
-			tmpSubArray := tmpDomain.([]int32)
-			tmpDimension := []int32{tmpSubArray[(2 * i)], tmpSubArray[(2*i)+1]}
-			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDimension)
-		case reflect.Int64:
-			tmpSubArray := tmpDomain.([]int64)
-			tmpDimension := []int64{tmpSubArray[(2 * i)], tmpSubArray[(2*i)+1]}
-			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDimension)
-		case reflect.Uint:
-			tmpSubArray := tmpDomain.([]uint)
-			tmpDimension := []uint{tmpSubArray[(2 * i)], tmpSubArray[(2*i)+1]}
-			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDimension)
-		case reflect.Uint8:
-			tmpSubArray := tmpDomain.([]uint8)
-			tmpDimension := []uint8{tmpSubArray[(2 * i)], tmpSubArray[(2*i)+1]}
-			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDimension)
-		case reflect.Uint16:
-			tmpSubArray := tmpDomain.([]uint16)
-			tmpDimension := []uint16{tmpSubArray[(2 * i)], tmpSubArray[(2*i)+1]}
-			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDimension)
-		case reflect.Uint32:
-			tmpSubArray := tmpDomain.([]uint32)
-			tmpDimension := []uint32{tmpSubArray[(2 * i)], tmpSubArray[(2*i)+1]}
-			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDimension)
-		case reflect.Uint64:
-			tmpSubArray := tmpDomain.([]uint64)
-			tmpDimension := []uint64{tmpSubArray[(2 * i)], tmpSubArray[(2*i)+1]}
-			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDimension)
-		case reflect.Float32:
-			tmpSubArray := tmpDomain.([]float32)
-			tmpDimension := []float32{tmpSubArray[(2 * i)], tmpSubArray[(2*i)+1]}
-			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDimension)
-		case reflect.Float64:
-			tmpSubArray := tmpDomain.([]float64)
-			tmpDimension := []float64{tmpSubArray[(2 * i)], tmpSubArray[(2*i)+1]}
-			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDimension)
-		default:
-			return nil, false, fmt.Errorf("unhandled domain datatype: %s", tmpDomainType.String())
+		switch tmpDomain := tmpDomain.(type) {
+		case []int:
+			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDomain[2*i:2*i+2])
+		case []int8:
+			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDomain[2*i:2*i+2])
+		case []int16:
+			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDomain[2*i:2*i+2])
+		case []int32:
+			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDomain[2*i:2*i+2])
+		case []int64:
+			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDomain[2*i:2*i+2])
+		case []uint:
+			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDomain[2*i:2*i+2])
+		case []uint8:
+			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDomain[2*i:2*i+2])
+		case []uint16:
+			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDomain[2*i:2*i+2])
+		case []uint32:
+			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDomain[2*i:2*i+2])
+		case []uint64:
+			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDomain[2*i:2*i+2])
+		case []float32:
+			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDomain[2*i:2*i+2])
+		case []float64:
+			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDomain[2*i:2*i+2])
+		case []bool:
+			nonEmptyDomain, err = getNonEmptyDomainForDim(dimension, tmpDomain[2*i:2*i+2])
 		}
 
 		if err != nil {
 			return nil, false, err
 		}
-		nonEmptyDomains = append(nonEmptyDomains, *nonEmptyDomain)
+		nonEmptyDomains[i] = *nonEmptyDomain
 	}
 
 	return nonEmptyDomains, false, nil
 }
 
-// SerializeArrayNonEmptyDomainAllDimensions gets and serializes the array nonempty domain
+// SerializeArrayNonEmptyDomainAllDimensions gets and serializes the array nonempty domain.
 func SerializeArrayNonEmptyDomainAllDimensions(a *Array, serializationType SerializationType) ([]byte, error) {
 
 	buffer := Buffer{context: a.context}
-	// Set finalizer for free C pointer on gc
-	runtime.SetFinalizer(&buffer, func(buffer *Buffer) {
-		buffer.Free()
-	})
+	freeOnGC(&buffer)
 
 	var cClientSide = C.int32_t(0) // Currently this parameter is unused in libtiledb
 	ret := C.tiledb_serialize_array_non_empty_domain_all_dimensions(a.context.tiledbContext, a.tiledbArray, C.tiledb_serialization_type_t(serializationType), cClientSide, &buffer.tiledbBuffer)
@@ -234,7 +194,7 @@ func SerializeArrayNonEmptyDomainAllDimensions(a *Array, serializationType Seria
 	return buffer.Serialize(serializationType)
 }
 
-// DeserializeArrayNonEmptyDomainAllDimensions deserializes an array nonempty domain
+// DeserializeArrayNonEmptyDomainAllDimensions deserializes an array nonempty domain.
 func DeserializeArrayNonEmptyDomainAllDimensions(a *Array, buffer *Buffer, serializationType SerializationType) error {
 
 	var cClientSide = C.int32_t(0) // Currently this parameter is unused in libtiledb
@@ -246,76 +206,10 @@ func DeserializeArrayNonEmptyDomainAllDimensions(a *Array, buffer *Buffer, seria
 	return nil
 }
 
-// SerializeArrayMaxBufferSizes gets and serializes the array max buffer sizes for the given subarray
-func SerializeArrayMaxBufferSizes(a *Array, subarray interface{}, serializationType SerializationType) ([]byte, error) {
-	// Create subarray void*
-	var cSubarray unsafe.Pointer
-	if reflect.TypeOf(subarray).Kind() != reflect.Slice {
-		return nil, fmt.Errorf("subarray passed must be a slice, type passed was: %s", reflect.TypeOf(subarray).Kind().String())
-	}
-	subarrayType := reflect.TypeOf(subarray).Elem().Kind()
-	switch subarrayType {
-	case reflect.Int:
-		tmpSubArray := subarray.([]int)
-		cSubarray = unsafe.Pointer(&tmpSubArray[0])
-	case reflect.Int8:
-		tmpSubArray := subarray.([]int8)
-		cSubarray = unsafe.Pointer(&tmpSubArray[0])
-	case reflect.Int16:
-		tmpSubArray := subarray.([]int16)
-		cSubarray = unsafe.Pointer(&tmpSubArray[0])
-	case reflect.Int32:
-		tmpSubArray := subarray.([]int32)
-		cSubarray = unsafe.Pointer(&tmpSubArray[0])
-	case reflect.Int64:
-		tmpSubArray := subarray.([]int64)
-		cSubarray = unsafe.Pointer(&tmpSubArray[0])
-	case reflect.Uint:
-		tmpSubArray := subarray.([]uint)
-		cSubarray = unsafe.Pointer(&tmpSubArray[0])
-	case reflect.Uint8:
-		tmpSubArray := subarray.([]uint8)
-		cSubarray = unsafe.Pointer(&tmpSubArray[0])
-	case reflect.Uint16:
-		tmpSubArray := subarray.([]uint16)
-		cSubarray = unsafe.Pointer(&tmpSubArray[0])
-	case reflect.Uint32:
-		tmpSubArray := subarray.([]uint32)
-		cSubarray = unsafe.Pointer(&tmpSubArray[0])
-	case reflect.Uint64:
-		tmpSubArray := subarray.([]uint64)
-		cSubarray = unsafe.Pointer(&tmpSubArray[0])
-	case reflect.Float32:
-		tmpSubArray := subarray.([]float32)
-		cSubarray = unsafe.Pointer(&tmpSubArray[0])
-	case reflect.Float64:
-		tmpSubArray := subarray.([]float64)
-		cSubarray = unsafe.Pointer(&tmpSubArray[0])
-	default:
-		return nil, fmt.Errorf("unhandled subarray datatype: %s", subarrayType.String())
-	}
-
-	buffer := Buffer{context: a.context}
-	// Set finalizer for free C pointer on gc
-	runtime.SetFinalizer(&buffer, func(buffer *Buffer) {
-		buffer.Free()
-	})
-
-	ret := C.tiledb_serialize_array_max_buffer_sizes(a.context.tiledbContext, a.tiledbArray, cSubarray, C.tiledb_serialization_type_t(serializationType), &buffer.tiledbBuffer)
-	if ret != C.TILEDB_OK {
-		return nil, fmt.Errorf("error serializing array max buffer sizes: %s", a.context.LastError())
-	}
-
-	return buffer.Serialize(serializationType)
-}
-
-// SerializeQuery serializes a query
+// SerializeQuery serializes a query.
 func SerializeQuery(query *Query, serializationType SerializationType, clientSide bool) (*BufferList, error) {
 	bufferList := BufferList{context: query.context}
-	// Set finalizer for free C pointer on gc
-	runtime.SetFinalizer(&bufferList, func(bufferList *BufferList) {
-		bufferList.Free()
-	})
+	freeOnGC(&bufferList)
 
 	var cClientSide C.int32_t
 	if clientSide {
@@ -332,7 +226,7 @@ func SerializeQuery(query *Query, serializationType SerializationType, clientSid
 	return &bufferList, nil
 }
 
-// DeserializeQuery deserializes a buffer into an existing query
+// DeserializeQuery deserializes a buffer into an existing query.
 func DeserializeQuery(query *Query, buffer *Buffer, serializationType SerializationType, clientSide bool) error {
 	var cClientSide C.int32_t
 	if clientSide {
@@ -349,13 +243,10 @@ func DeserializeQuery(query *Query, buffer *Buffer, serializationType Serializat
 	return nil
 }
 
-// SerializeArrayMetadata gets and serializes the array metadata
+// SerializeArrayMetadata gets and serializes the array metadata.
 func SerializeArrayMetadata(a *Array, serializationType SerializationType) ([]byte, error) {
 	buffer := Buffer{context: a.context}
-	// Set finalizer for free C pointer on gc
-	runtime.SetFinalizer(&buffer, func(buffer *Buffer) {
-		buffer.Free()
-	})
+	freeOnGC(&buffer)
 
 	ret := C.tiledb_serialize_array_metadata(a.context.tiledbContext, a.tiledbArray, C.tiledb_serialization_type_t(serializationType), &buffer.tiledbBuffer)
 	if ret != C.TILEDB_OK {
@@ -365,7 +256,7 @@ func SerializeArrayMetadata(a *Array, serializationType SerializationType) ([]by
 	return buffer.Serialize(serializationType)
 }
 
-// DeserializeArrayMetadata deserializes array metadata
+// DeserializeArrayMetadata deserializes array metadata.
 func DeserializeArrayMetadata(a *Array, buffer *Buffer, serializationType SerializationType) error {
 	ret := C.tiledb_deserialize_array_metadata(a.context.tiledbContext, a.tiledbArray, C.tiledb_serialization_type_t(serializationType), buffer.tiledbBuffer)
 	if ret != C.TILEDB_OK {
@@ -374,7 +265,7 @@ func DeserializeArrayMetadata(a *Array, buffer *Buffer, serializationType Serial
 	return nil
 }
 
-// SerializeQueryEstResultSizes gets and serializes the query estimated result sizes
+// SerializeQueryEstResultSizes gets and serializes the query estimated result sizes.
 func SerializeQueryEstResultSizes(q *Query, serializationType SerializationType, clientSide bool) ([]byte, error) {
 	var cClientSide C.int32_t
 	if clientSide {
@@ -384,10 +275,7 @@ func SerializeQueryEstResultSizes(q *Query, serializationType SerializationType,
 	}
 
 	buffer := Buffer{context: q.context}
-	// Set finalizer for free C pointer on gc
-	runtime.SetFinalizer(&buffer, func(buffer *Buffer) {
-		buffer.Free()
-	})
+	freeOnGC(&buffer)
 
 	ret := C.tiledb_serialize_query_est_result_sizes(q.context.tiledbContext, q.tiledbQuery, C.tiledb_serialization_type_t(serializationType), cClientSide, &buffer.tiledbBuffer)
 	if ret != C.TILEDB_OK {
@@ -397,7 +285,7 @@ func SerializeQueryEstResultSizes(q *Query, serializationType SerializationType,
 	return buffer.Serialize(serializationType)
 }
 
-// DeserializeQueryEstResultSizes deserializes query estimated result sizes
+// DeserializeQueryEstResultSizes deserializes query estimated result sizes.
 func DeserializeQueryEstResultSizes(q *Query, buffer *Buffer, serializationType SerializationType, clientSide bool) error {
 	var cClientSide C.int32_t
 	if clientSide {
@@ -411,4 +299,253 @@ func DeserializeQueryEstResultSizes(q *Query, buffer *Buffer, serializationType 
 		return fmt.Errorf("Error deserializing query est buffer sizes: %s", q.context.LastError())
 	}
 	return nil
+}
+
+// SerializeArray serializes an array.
+func SerializeArray(array *Array, serializationType SerializationType, clientSide bool) ([]byte, error) {
+	var cClientSide C.int32_t
+	if clientSide {
+		cClientSide = 1
+	} else {
+		cClientSide = 0
+	}
+
+	buffer := Buffer{context: array.context}
+	// Set finalizer for free C pointer on gc
+	freeOnGC(&buffer)
+
+	ret := C.tiledb_serialize_array(array.context.tiledbContext, array.tiledbArray, C.tiledb_serialization_type_t(serializationType), cClientSide, &buffer.tiledbBuffer)
+	if ret != C.TILEDB_OK {
+		return nil, fmt.Errorf("error serializing array: %s", array.context.LastError())
+	}
+
+	return buffer.Serialize(serializationType)
+}
+
+// DeserializeArray deserializes a new array from the given buffer.
+func DeserializeArray(buffer *Buffer, serializationType SerializationType, clientSide bool, arrayURI string) (*Array, error) {
+	array := Array{context: buffer.context}
+
+	var cClientSide C.int32_t
+	if clientSide {
+		cClientSide = 1
+	} else {
+		cClientSide = 0
+	}
+
+	cArrayURI := C.CString(arrayURI)
+	defer C.free(unsafe.Pointer(cArrayURI))
+
+	ret := C.tiledb_deserialize_array(array.context.tiledbContext, buffer.tiledbBuffer, C.tiledb_serialization_type_t(serializationType), cClientSide, cArrayURI, &array.tiledbArray)
+	if ret != C.TILEDB_OK {
+		return nil, fmt.Errorf("error deserializing array: %s", array.context.LastError())
+	}
+
+	// Set finalizer for free C pointer on gc
+	// This needs to happen *after* the tiledb_deserialize_array call
+	// because that may leave the array with a non-nil pointer
+	// to already-freed memory.
+	freeOnGC(&array)
+
+	return &array, nil
+}
+
+// SerializeFragmentInfo serializes fragment info.
+func SerializeFragmentInfo(fragmentInfo *FragmentInfo, serializationType SerializationType, clientSide bool) ([]byte, error) {
+	var cClientSide C.int32_t
+	if clientSide {
+		cClientSide = 1
+	} else {
+		cClientSide = 0
+	}
+
+	buffer := Buffer{context: fragmentInfo.context}
+	// Set finalizer for free C pointer on gc
+	freeOnGC(&buffer)
+
+	ret := C.tiledb_serialize_fragment_info(fragmentInfo.context.tiledbContext, fragmentInfo.tiledbFragmentInfo, C.tiledb_serialization_type_t(serializationType), cClientSide, &buffer.tiledbBuffer)
+	if ret != C.TILEDB_OK {
+		return nil, fmt.Errorf("error serializing array: %s", fragmentInfo.context.LastError())
+	}
+
+	return buffer.Serialize(serializationType)
+}
+
+// DeserializeFragmentInfo deserializes an existing fragment info from the given buffer.
+func DeserializeFragmentInfo(fragmentInfo FragmentInfo, buffer *Buffer, arrayURI string, serializationType SerializationType, clientSide bool) error {
+	var cClientSide C.int32_t
+	if clientSide {
+		cClientSide = 1
+	} else {
+		cClientSide = 0
+	}
+
+	cArrayURI := C.CString(arrayURI)
+	defer C.free(unsafe.Pointer(cArrayURI))
+
+	ret := C.tiledb_deserialize_fragment_info(fragmentInfo.context.tiledbContext, buffer.tiledbBuffer, C.tiledb_serialization_type_t(serializationType), cArrayURI, cClientSide, fragmentInfo.tiledbFragmentInfo)
+	if ret != C.TILEDB_OK {
+		return fmt.Errorf("error deserializing array: %s", fragmentInfo.context.LastError())
+	}
+
+	return nil
+}
+
+// SerializeFragmentInfoRequest serializes fragment info.
+func SerializeFragmentInfoRequest(fragmentInfo *FragmentInfo, serializationType SerializationType, clientSide bool) ([]byte, error) {
+	var cClientSide C.int32_t
+	if clientSide {
+		cClientSide = 1
+	} else {
+		cClientSide = 0
+	}
+
+	buffer := Buffer{context: fragmentInfo.context}
+	// Set finalizer for free C pointer on gc
+	freeOnGC(&buffer)
+
+	ret := C.tiledb_serialize_fragment_info_request(fragmentInfo.context.tiledbContext, fragmentInfo.tiledbFragmentInfo, C.tiledb_serialization_type_t(serializationType), cClientSide, &buffer.tiledbBuffer)
+	if ret != C.TILEDB_OK {
+		return nil, fmt.Errorf("error serializing array: %s", fragmentInfo.context.LastError())
+	}
+
+	return buffer.Serialize(serializationType)
+}
+
+// DeserializeFragmentInfoRequest deserializes an existing fragment info from the given buffer.
+func DeserializeFragmentInfoRequest(fragmentInfo FragmentInfo, buffer *Buffer, serializationType SerializationType, clientSide bool) error {
+	var cClientSide C.int32_t
+	if clientSide {
+		cClientSide = 1
+	} else {
+		cClientSide = 0
+	}
+
+	ret := C.tiledb_deserialize_fragment_info_request(fragmentInfo.context.tiledbContext, buffer.tiledbBuffer, C.tiledb_serialization_type_t(serializationType), cClientSide, fragmentInfo.tiledbFragmentInfo)
+	if ret != C.TILEDB_OK {
+		return fmt.Errorf("error deserializing array: %s", fragmentInfo.context.LastError())
+	}
+
+	return nil
+}
+
+func DeserializeQueryAndArray(context *Context, buffer *Buffer, serializationType SerializationType, clientSide bool, arrayURI string) (*Array, *Query, error) {
+	var cClientSide C.int32_t
+	if clientSide {
+		cClientSide = 1
+	} else {
+		cClientSide = 0
+	}
+
+	cArrayURI := C.CString(arrayURI)
+	defer C.free(unsafe.Pointer(cArrayURI))
+
+	array := &Array{
+		context: context,
+	}
+
+	query := &Query{
+		context: context,
+		array:   array,
+	}
+
+	ret := C.tiledb_deserialize_query_and_array(context.tiledbContext, buffer.tiledbBuffer, C.tiledb_serialization_type_t(serializationType), cClientSide, cArrayURI, &query.tiledbQuery, &array.tiledbArray)
+	if ret != C.TILEDB_OK {
+		return nil, nil, fmt.Errorf("error deserializing query: %s", context.LastError())
+	}
+
+	freeOnGC(array)
+	freeOnGC(query)
+
+	query.resultBufferElements = make(map[string][3]*uint64)
+
+	// Make sure the buffer stays alive untill after the deserialization is complete
+	runtime.KeepAlive(buffer)
+	return array, query, nil
+}
+
+// HandleLoadArraySchemaRequest Passes the array and serialized LoadArraySchemaRequest to core which returns the
+// serialized LoadArraySchemaResponse. The request contains a TileDB Config used to load the schema, the response
+// contains the latest array schema loaded and a map of all array schemas.
+func HandleLoadArraySchemaRequest(array *Array, request *Buffer, serializationType SerializationType) (*Buffer, error) {
+	response, err := NewBuffer(array.context)
+	if err != nil {
+		return nil, fmt.Errorf("error creating LoadArraySchemaResponse buffer: %s", array.context.LastError())
+	}
+
+	ret := C.tiledb_handle_load_array_schema_request(array.context.tiledbContext, array.tiledbArray, C.tiledb_serialization_type_t(serializationType), request.tiledbBuffer, response.tiledbBuffer)
+	if ret != C.TILEDB_OK {
+		return nil, fmt.Errorf("error handling LoadArraySchemaRequset: %s", array.context.LastError())
+	}
+
+	runtime.KeepAlive(request)
+	return response, nil
+}
+
+// HandleArrayDeleteFragmentsTimestampsRequest is used by TileDB cloud to handle DeleteFragments with tiledb:// uris.
+func HandleArrayDeleteFragmentsTimestampsRequest(context *Context, array *Array, buffer *Buffer, serializationType SerializationType) error {
+	ret := C.tiledb_handle_array_delete_fragments_timestamps_request(context.tiledbContext, array.tiledbArray,
+		C.tiledb_serialization_type_t(serializationType), buffer.tiledbBuffer)
+	if ret != C.TILEDB_OK {
+		return fmt.Errorf("error deserializing delete fragments timestamps: %s", context.LastError())
+	}
+
+	runtime.KeepAlive(buffer)
+	return nil
+}
+
+// HandleArrayDeleteFragmentsListRequest is used by TileDB cloud to handle DeleteFragmentsList with tiledb:// uris.
+func HandleArrayDeleteFragmentsListRequest(context *Context, array *Array, buffer *Buffer, serializationType SerializationType) error {
+	ret := C.tiledb_handle_array_delete_fragments_list_request(context.tiledbContext, array.tiledbArray,
+		C.tiledb_serialization_type_t(serializationType), buffer.tiledbBuffer)
+	if ret != C.TILEDB_OK {
+		return fmt.Errorf("error deserializing delete fragments list: %s", context.LastError())
+	}
+
+	runtime.KeepAlive(buffer)
+	return nil
+}
+
+// HandleQueryPlanRequest handles a request for a query plan. This is used by TileDB-Cloud
+// It returns a buffer with the serialized response. The caller should free the buffer after use.
+func HandleQueryPlanRequest(array *Array, serializationType SerializationType, request *Buffer) (*Buffer, error) {
+	opContext := array.context
+
+	response, err := NewBuffer(opContext)
+	if err != nil {
+		return nil, fmt.Errorf("Error allocating tiledb buffer: %s", opContext.LastError())
+	}
+
+	ret := C.tiledb_handle_query_plan_request(opContext.tiledbContext, array.tiledbArray, C.tiledb_serialization_type_t(serializationType),
+		request.tiledbBuffer, response.tiledbBuffer)
+	if ret != C.TILEDB_OK {
+		return nil, fmt.Errorf("Error handling query plan request: %s", opContext.LastError())
+	}
+
+	runtime.KeepAlive(request)
+	runtime.KeepAlive(array)
+
+	return response, nil
+}
+
+// HandleConsolidationPlanRequest handles a request for a consolidation plan. This is used by TileDB-Cloud
+// It returns a buffer with the serialized response. The caller should free the buffer after use.
+func HandleConsolidationPlanRequest(array *Array, serializationType SerializationType, request *Buffer) (*Buffer, error) {
+	opContext := array.context
+
+	response, err := NewBuffer(opContext)
+	if err != nil {
+		return nil, fmt.Errorf("Error allocating tiledb buffer: %s", opContext.LastError())
+	}
+
+	ret := C.tiledb_handle_consolidation_plan_request(opContext.tiledbContext, array.tiledbArray, C.tiledb_serialization_type_t(serializationType),
+		request.tiledbBuffer, response.tiledbBuffer)
+	if ret != C.TILEDB_OK {
+		return nil, fmt.Errorf("Error handling consolidation plan request: %s", opContext.LastError())
+	}
+
+	runtime.KeepAlive(request)
+	runtime.KeepAlive(array)
+
+	return response, nil
 }
